Module fred::clients

Redis client implementations.

Structs

ExclusivePool

A cheaply cloneable round-robin client pool that provides exclusive ownership over the inner clients.

Pipeline

Send a series of commands in a pipeline.

RedisClient

A cheaply cloneable Redis client struct.

RedisPool

A cheaply cloneable round-robin client pool.

Replicasreplicas

A struct for interacting with cluster replica nodes.

SentinelClientsentinel-client

A struct for interacting directly with Sentinel nodes.

SubscriberClientsubscriber-client

A subscriber client that will manage subscription state to any pubsub channels or patterns for the caller.

Transactiontransactions

A cheaply cloneable transaction block.

WithOptions

A client interface used to customize command configuration options.