Struct fred::clients::ExclusivePool

pub struct ExclusivePool { /* private fields */ }

A cheaply cloneable round-robin client pool that provides exclusive ownership over the inner clients.

This interface can be used when callers require exclusive ownership over the connection. For example,

WATCH foo
foo = GET foo
if foo > 1:
  MULTI
    INCR foo
    INCR bar
    INCR baz
  EXEC

Unlike RedisPool, this pooling interface does not directly implement ClientLike. Callers acquire and release clients via the returned MutexGuard.

use fred::{
  clients::{ExclusivePool, RedisPool},
  prelude::*,
};

async fn example() -> Result<(), RedisError> {
  let builder = Builder::default_centralized();
  let shared_pool = builder.build_pool(5)?;
  let exclusive_pool = builder.build_exclusive_pool(5)?;
  shared_pool.init().await?;
  exclusive_pool.init().await?;

  // since `RedisPool` implements `ClientLike` we can use most command interfaces directly
  let foo: Option<String> = shared_pool.set("foo", 1, None, None, false).await?;

  // with an `ExclusivePool` callers acquire and release clients with an async lock guard
  let results: Option<(i64, i64, i64)> = {
    let client = exclusive_pool.acquire().await;

    client.watch("foo").await?;
    if let Some(1) = client.get::<Option<i64>, _>("foo").await? {
      let trx = client.multi();
      trx.incr("foo").await?;
      trx.incr("bar").await?;
      trx.incr("baz").await?;
      Some(trx.exec(true).await?)
    } else {
      None
    }
  };
  assert_eq!(results, Some((2, 1, 1)));

  Ok(())
}

Callers should avoid cloning the inner clients, if possible.

Implementations

impl ExclusivePool

pub fn new( config: RedisConfig, perf: Option<PerformanceConfig>, connection: Option<ConnectionConfig>, policy: Option<ReconnectPolicy>, size: usize, ) -> Result<Self, RedisError>

Create a new pool without connecting to the server.

See the builder interface for more information.

pub fn clients(&self) -> &[Arc<AsyncMutex<RedisClient>>]

Read the clients in the pool.

pub async fn connect_pool(&self) -> Vec<ConnectHandle>

Connect each client to the server, returning the task driving each connection.

Use the base connect function to return one handle that drives all connections via join.

pub async fn connect(&self) -> ConnectHandle

Connect each client to the server.

This function returns a JoinHandle to a task that drives all connections via join.

See connect_pool for a variation of this function that separates the connection tasks.

See init for an alternative shorthand.

pub async fn force_reconnection(&self) -> RedisResult<()>

Force a reconnection to the server(s) for each client.

When running against a cluster this function will also refresh the cached cluster routing table.

pub async fn wait_for_connect(&self) -> RedisResult<()>

Wait for all the clients to connect to the server.

pub async fn init(&self) -> RedisResult<ConnectHandle>

Initialize a new routing and connection task for each client and wait for them to connect successfully.

The returned ConnectHandle refers to the task that drives the routing and connection layer for each client. It will not finish until the max reconnection count is reached.

Callers can also use connect and wait_for_connect separately if needed.

use fred::prelude::*;

#[tokio::main]
async fn main() -> Result<(), RedisError> {
  let pool = Builder::default_centralized().build_exclusive_pool(5)?;
  let connection_task = pool.init().await?;

  // ...

  pool.quit().await?;
  connection_task.await?
}

pub fn size(&self) -> usize

Read the size of the pool.

pub async fn acquire(&self) -> OwnedMutexGuard<RedisClient>

Read the client that should run the next command.

pub async fn update_perf_config(&self, config: PerformanceConfig)

Update the internal PerformanceConfig on each client in place with new values.

pub async fn set_resolver(&self, resolver: Arc<dyn Resolve>)

Available on crate feature dns only.

Override the DNS resolution logic for all clients in the pool.

pub async fn quit(&self) -> RedisResult<()>

Close the connection to the Redis server for each client. The returned future resolves when the command has been written to the socket, not when the connection has been fully closed. Some time after this future resolves the future returned by connect will resolve which indicates that the connection has been fully closed.

This function will also close all error, pubsub message, and reconnection event streams on all clients in the pool.