Struct fred::clients::RedisPool

pub struct RedisPool { /* private fields */ }

A cheaply cloneable round-robin client pool.

Restrictions

The following interfaces are not implemented on RedisPool:

MetricsInterface

PubsubInterface

EventInterface

ClientInterface

AuthInterface

In many cases, such as publish, callers can work around this by adding a call to next, but in some scenarios this may not work. As a general rule, any commands that change or depend on local connection state will not be implemented directly on RedisPool. Callers can use clients, next, or last to operate on individual clients if needed.

Implementations

impl RedisPool

pub fn from_clients(clients: Vec<RedisClient>) -> Result<Self, RedisError>

Create a new pool from an existing set of clients.

pub fn new( config: RedisConfig, perf: Option<PerformanceConfig>, connection: Option<ConnectionConfig>, policy: Option<ReconnectPolicy>, size: usize, ) -> Result<Self, RedisError>

Create a new pool without connecting to the server.

See the builder interface for more information.

pub fn prefer_connected(&self, val: bool) -> bool

Set whether the client will use next_connected or next when routing commands among the pooled clients.

pub fn clients(&self) -> &[RedisClient]

Read the individual clients in the pool.

pub fn connect_pool(&self) -> Vec<ConnectHandle>

Connect each client to the server, returning the task driving each connection.

Use the base connect function to return one handle that drives all connections via join.

pub fn size(&self) -> usize

Read the size of the pool.

pub fn next_connected(&self) -> &RedisClient

Read the next connected client that should run the next command.

pub fn next(&self) -> &RedisClient

Read the client that should run the next command.

pub fn last(&self) -> &RedisClient

Read the client that ran the last command.

pub fn replicas(&self) -> Replicas

Available on crate feature replicas only.

Create a client that interacts with the replica nodes associated with the next client.