Struct fred::clients::SentinelClient

pub struct SentinelClient { /* private fields */ }

Available on crate feature sentinel-client only.

A struct for interacting directly with Sentinel nodes.

This struct will not communicate with Redis servers behind the sentinel interface, but rather with the sentinel nodes themselves. Callers should use the RedisClient interface with a ServerConfig::Sentinel for interacting with Redis services behind a sentinel layer.

See the sentinel API docs for more information.

Implementations

impl SentinelClient

pub fn new( config: SentinelConfig, perf: Option<PerformanceConfig>, connection: Option<ConnectionConfig>, policy: Option<ReconnectPolicy>, ) -> SentinelClient

Create a new client instance without connecting to the sentinel node.

See the builder interface for more information.