Enum fred::error::RedisErrorKind

pub enum RedisErrorKind {
    Auth,
    IO(IoError),
    InvalidCommand,
    InvalidArgument,
    UrlError(ParseError),
    ProtocolError,
    Canceled,
    Unknown,
    // some variants omitted
}

An enum representing the type of error from Redis.

Variants
Auth

An authentication error.

IO(IoError)

An IO error with the underlying connection.

InvalidCommand

An invalid command, such as trying to perform a set command on a client after calling subscribe.

InvalidArgument

An invalid argument or set of arguments to a command.

UrlError(ParseError)

An invalid URL error.

ProtocolError

A protocol error.

Canceled

An error indicating the request was canceled.

Unknown

An unknown error.