Module fred::interfaces

Traits that implement portions of the Redis interface.

Enums

Resp3Frame

A RESP3 frame that uses Bytes and Str as the underlying buffer type.

Traits

AclInterfacei-acl

Functions that implement the ACL interface.

AuthInterface

Functions for authenticating clients.

ClientInterfacei-client

Functions that implement the client interface.

ClientLike

Any Redis client that implements any part of the Redis interface.

ClusterInterfacei-cluster

Functions that implement the cluster interface.

ConfigInterfacei-config

Functions that implement the config interface.

EventInterface

An interface that exposes various client and connection events.

FunctionInterfacei-scripts

Functions that implement the function interface.

GeoInterfacei-geo

Functions that implement the geo interface.

HashesInterfacei-hashes

Functions that implement the hashes interface.

HeartbeatInterface

Functions that provide a connection heartbeat interface.

HyperloglogInterfacei-hyperloglog

Functions that implement the HyperLogLog interface.

KeysInterfacei-keys

Functions that implement the generic keys interface.

ListInterfacei-lists

Functions that implement the lists interface.

LuaInterfacei-scripts

Functions that implement the lua interface.

MemoryInterfacei-memory

Functions that implement the memory interface.

MetricsInterface

Functions that implement the internal metrics interface.

PubsubInterfacei-pubsub

Functions that implement the pubsub interface.

RediSearchInterfacei-redisearch

A RediSearch interface.

RedisJsonInterfacei-redis-json

The client commands in the RedisJSON interface.

SentinelInterfacesentinel-client

Functions that implement the sentinel interface.

ServerInterfacei-server

Functions that implement the server interface.

SetsInterfacei-sets

Functions that implement the sets interface.

SlowlogInterfacei-slowlog

Functions that implement the slowlog interface.

SortedSetsInterfacei-sorted-sets

Functions that implement the sorted sets interface.

StreamsInterfacei-streams

Functions that implement the streams interface.

TimeSeriesInterfacei-time-series

A Redis Timeseries interface.

TrackingInterfacei-tracking

A high level interface that supports client side caching via the client tracking interface.

TransactionInterfacetransactions

Functions that implement the transactions interface.

Functions

spawn_event_listener

Type Aliases

RedisResult

Type alias for Result<T, RedisError>.