Trait fred::interfaces::AclInterface

pub trait AclInterface: ClientLike + Sized {
    // Provided methods
    fn acl_setuser<S, V>(
        &self,
        username: S,
        rules: V,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<Str> + Send,
             V: TryInto<MultipleValues> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn acl_load(&self) -> impl Future<Output = RedisResult<()>> + Send { ... }
    fn acl_save(&self) -> impl Future<Output = RedisResult<()>> + Send { ... }
    fn acl_list<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn acl_users<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn acl_getuser<R, S>(
        &self,
        username: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn acl_deluser<R, S>(
        &self,
        usernames: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<MultipleStrings> + Send { ... }
    fn acl_cat<R>(
        &self,
        category: Option<Str>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn acl_genpass<R>(
        &self,
        bits: Option<u16>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn acl_whoami<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn acl_log_count<R>(
        &self,
        count: Option<u32>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn acl_log_reset(&self) -> impl Future<Output = RedisResult<()>> + Send { ... }
}

Available on crate feature i-acl only.

Functions that implement the ACL interface.

Provided Methods

fn acl_setuser<S, V>( &self, username: S, rules: V, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<Str> + Send, V: TryInto<MultipleValues> + Send, V::Error: Into<RedisError> + Send,

Create an ACL user with the specified rules or modify the rules of an existing user.

https://redis.io/commands/acl-setuser

fn acl_load(&self) -> impl Future<Output = RedisResult<()>> + Send

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will reload the ACLs from the file, replacing all the current ACL rules with the ones defined in the file.

https://redis.io/commands/acl-load

fn acl_save(&self) -> impl Future<Output = RedisResult<()>> + Send

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will save the currently defined ACLs from the server memory to the ACL file.

https://redis.io/commands/acl-save

fn acl_list<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

The command shows the currently active ACL rules in the Redis server.

https://redis.io/commands/acl-list\

fn acl_users<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

The command shows a list of all the usernames of the currently configured users in the Redis ACL system.

https://redis.io/commands/acl-users

fn acl_getuser<R, S>( &self, username: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

The command returns all the rules defined for an existing ACL user.

https://redis.io/commands/acl-getuser

fn acl_deluser<R, S>( &self, usernames: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<MultipleStrings> + Send,

Delete all the specified ACL users and terminate all the connections that are authenticated with such users.

https://redis.io/commands/acl-deluser

fn acl_cat<R>( &self, category: Option<Str>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

The command shows the available ACL categories if called without arguments. If a category name is given, the command shows all the Redis commands in the specified category.

https://redis.io/commands/acl-cat

fn acl_genpass<R>( &self, bits: Option<u16>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Generate a password with length bits, returning the password.

https://redis.io/commands/acl-genpass

fn acl_whoami<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Return the username the current connection is authenticated with. New connections are authenticated with the “default” user.

https://redis.io/commands/acl-whoami

fn acl_log_count<R>( &self, count: Option<u32>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Read count recent ACL security events.

https://redis.io/commands/acl-log

fn acl_log_reset(&self) -> impl Future<Output = RedisResult<()>> + Send

Clear the ACL security events logs.

https://redis.io/commands/acl-log

Object Safety

This trait is not object safe.