Trait fred::interfaces::AuthInterface

pub trait AuthInterface: ClientLike {
    // Provided methods
    fn auth<S>(
        &self,
        username: Option<String>,
        password: S,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<Str> + Send { ... }
    fn hello(
        &self,
        version: RespVersion,
        auth: Option<(Str, Str)>,
        setname: Option<Str>,
    ) -> impl Future<Output = RedisResult<()>> + Send { ... }
}

Functions for authenticating clients.

Provided Methods

fn auth<S>( &self, username: Option<String>, password: S, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<Str> + Send,

Request for authentication in a password-protected Redis server. Returns ok if successful.

The client will automatically authenticate with the default user if a password is provided in the associated RedisConfig when calling connect.

If running against clustered servers this function will authenticate all connections.

https://redis.io/commands/auth

fn hello( &self, version: RespVersion, auth: Option<(Str, Str)>, setname: Option<Str>, ) -> impl Future<Output = RedisResult<()>> + Send

Switch to a different protocol, optionally authenticating in the process.

If running against clustered servers this function will issue the HELLO command to each server concurrently.

https://redis.io/commands/hello

Object Safety

This trait is not object safe.