Trait fred::interfaces::FunctionInterface

pub trait FunctionInterface: ClientLike + Sized {

    // Provided methods
    fn fcall<R, F, K, V>(
        &self,
        func: F,
        keys: K,
        args: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             F: Into<Str> + Send,
             K: Into<MultipleKeys> + Send,
             V: TryInto<MultipleValues> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn fcall_ro<R, F, K, V>(
        &self,
        func: F,
        keys: K,
        args: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             F: Into<Str> + Send,
             K: Into<MultipleKeys> + Send,
             V: TryInto<MultipleValues> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn function_delete<R, S>(
        &self,
        library_name: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn function_delete_cluster<S>(
        &self,
        library_name: S,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<Str> + Send { ... }
    fn function_dump<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn function_flush<R>(
        &self,
        async: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn function_flush_cluster(
        &self,
        async: bool,
    ) -> impl Future<Output = RedisResult<()>> + Send { ... }
    fn function_kill<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn function_list<R, S>(
        &self,
        library_name: Option<S>,
        withcode: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn function_load<R, S>(
        &self,
        replace: bool,
        code: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn function_load_cluster<R, S>(
        &self,
        replace: bool,
        code: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn function_restore<R, B, P>(
        &self,
        serialized: B,
        policy: P,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             B: Into<Bytes> + Send,
             P: TryInto<FnPolicy> + Send,
             P::Error: Into<RedisError> + Send { ... }
    fn function_restore_cluster<B, P>(
        &self,
        serialized: B,
        policy: P,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where B: Into<Bytes> + Send,
             P: TryInto<FnPolicy> + Send,
             P::Error: Into<RedisError> + Send { ... }
    fn function_stats<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }

}

Available on crate feature i-scripts only.

Functions that implement the function interface.

Provided Methods

fn fcall<R, F, K, V>( &self, func: F, keys: K, args: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, F: Into<Str> + Send, K: Into<MultipleKeys> + Send, V: TryInto<MultipleValues> + Send, V::Error: Into<RedisError> + Send,

Invoke a function.

https://redis.io/commands/fcall/

fn fcall_ro<R, F, K, V>( &self, func: F, keys: K, args: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, F: Into<Str> + Send, K: Into<MultipleKeys> + Send, V: TryInto<MultipleValues> + Send, V::Error: Into<RedisError> + Send,

This is a read-only variant of the FCALL command that cannot execute commands that modify data.

https://redis.io/commands/fcall_ro/

fn function_delete<R, S>( &self, library_name: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

Delete a library and all its functions.

https://redis.io/commands/function-delete/

fn function_delete_cluster<S>( &self, library_name: S, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<Str> + Send,

Delete a library and all its functions from each cluster node concurrently.

https://redis.io/commands/function-delete/

fn function_dump<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Return the serialized payload of loaded libraries.

https://redis.io/commands/function-dump/

fn function_flush<R>( &self, async: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Deletes all the libraries.

https://redis.io/commands/function-flush/

fn function_flush_cluster( &self, async: bool, ) -> impl Future<Output = RedisResult<()>> + Send

Deletes all the libraries on all cluster nodes concurrently.

https://redis.io/commands/function-flush/

fn function_kill<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Kill a function that is currently executing.

Note: This command runs on a backchannel connection to the server in order to take effect as quickly as possible.

https://redis.io/commands/function-kill/

fn function_list<R, S>( &self, library_name: Option<S>, withcode: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

Return information about the functions and libraries.

https://redis.io/commands/function-list/

fn function_load<R, S>( &self, replace: bool, code: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

Load a library to Redis.

https://redis.io/commands/function-load/

fn function_load_cluster<R, S>( &self, replace: bool, code: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

Load a library to Redis on all cluster nodes concurrently.

https://redis.io/commands/function-load/

fn function_restore<R, B, P>( &self, serialized: B, policy: P, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, B: Into<Bytes> + Send, P: TryInto<FnPolicy> + Send, P::Error: Into<RedisError> + Send,

Restore libraries from the serialized payload.

https://redis.io/commands/function-restore/

Note: Use FnPolicy::default() to use the default function restore policy ("APPEND").

fn function_restore_cluster<B, P>( &self, serialized: B, policy: P, ) -> impl Future<Output = RedisResult<()>> + Send
where B: Into<Bytes> + Send, P: TryInto<FnPolicy> + Send, P::Error: Into<RedisError> + Send,

Restore libraries from the serialized payload on all cluster nodes concurrently.

https://redis.io/commands/function-restore/

Note: Use FnPolicy::default() to use the default function restore policy ("APPEND").

fn function_stats<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Return information about the function that’s currently running and information about the available execution engines.

Note: This command runs on a backchannel connection to the server.

https://redis.io/commands/function-stats/

Object Safety

This trait is not object safe.