Trait fred::interfaces::HashesInterface

pub trait HashesInterface: ClientLike + Sized {

    // Provided methods
    fn hgetall<R, K>(
        &self,
        key: K,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn hdel<R, K, F>(
        &self,
        key: K,
        fields: F,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<MultipleKeys> + Send { ... }
    fn hexists<R, K, F>(
        &self,
        key: K,
        field: F,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<RedisKey> + Send { ... }
    fn hget<R, K, F>(
        &self,
        key: K,
        field: F,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<RedisKey> + Send { ... }
    fn hincrby<R, K, F>(
        &self,
        key: K,
        field: F,
        increment: i64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<RedisKey> + Send { ... }
    fn hincrbyfloat<R, K, F>(
        &self,
        key: K,
        field: F,
        increment: f64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<RedisKey> + Send { ... }
    fn hkeys<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn hlen<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn hmget<R, K, F>(
        &self,
        key: K,
        fields: F,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<MultipleKeys> + Send { ... }
    fn hmset<R, K, V>(
        &self,
        key: K,
        values: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<RedisMap> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn hset<R, K, V>(
        &self,
        key: K,
        values: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<RedisMap> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn hsetnx<R, K, F, V>(
        &self,
        key: K,
        field: F,
        value: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<RedisKey> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn hrandfield<R, K>(
        &self,
        key: K,
        count: Option<(i64, bool)>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn hstrlen<R, K, F>(
        &self,
        key: K,
        field: F,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: Into<RedisKey> + Send { ... }
    fn hvals<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }

}

Available on crate feature i-hashes only.

Functions that implement the hashes interface.

Provided Methods

fn hgetall<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns all fields and values of the hash stored at key.

https://redis.io/commands/hgetall

fn hdel<R, K, F>( &self, key: K, fields: F, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<MultipleKeys> + Send,

Removes the specified fields from the hash stored at key.

https://redis.io/commands/hdel

fn hexists<R, K, F>( &self, key: K, field: F, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<RedisKey> + Send,

Returns if field is an existing field in the hash stored at key.

https://redis.io/commands/hexists

fn hget<R, K, F>( &self, key: K, field: F, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<RedisKey> + Send,

Returns the value associated with field in the hash stored at key.

https://redis.io/commands/hget

fn hincrby<R, K, F>( &self, key: K, field: F, increment: i64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<RedisKey> + Send,

Increments the number stored at field in the hash stored at key by increment.

https://redis.io/commands/hincrby

fn hincrbyfloat<R, K, F>( &self, key: K, field: F, increment: f64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<RedisKey> + Send,

Increment the specified field of a hash stored at key, and representing a floating point number, by the specified increment.

https://redis.io/commands/hincrbyfloat

fn hkeys<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns all field names in the hash stored at key.

https://redis.io/commands/hkeys

fn hlen<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns the number of fields contained in the hash stored at key.

https://redis.io/commands/hlen

fn hmget<R, K, F>( &self, key: K, fields: F, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<MultipleKeys> + Send,

Returns the values associated with the specified fields in the hash stored at key.

https://redis.io/commands/hmget

fn hmset<R, K, V>( &self, key: K, values: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<RedisMap> + Send, V::Error: Into<RedisError> + Send,

Sets the specified fields to their respective values in the hash stored at key.

https://redis.io/commands/hmset

fn hset<R, K, V>( &self, key: K, values: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<RedisMap> + Send, V::Error: Into<RedisError> + Send,

Sets fields in the hash stored at key to their provided values.

https://redis.io/commands/hset

fn hsetnx<R, K, F, V>( &self, key: K, field: F, value: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<RedisKey> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Sets field in the hash stored at key to value, only if field does not yet exist.

https://redis.io/commands/hsetnx

fn hrandfield<R, K>( &self, key: K, count: Option<(i64, bool)>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

When called with just the key argument, return a random field from the hash value stored at key.

If the provided count argument is positive, return an array of distinct fields.

https://redis.io/commands/hrandfield

fn hstrlen<R, K, F>( &self, key: K, field: F, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: Into<RedisKey> + Send,

Returns the string length of the value associated with field in the hash stored at key.

https://redis.io/commands/hstrlen

fn hvals<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns all values in the hash stored at key.

https://redis.io/commands/hvals

Object Safety

This trait is not object safe.