Trait fred::interfaces::MemoryInterface

pub trait MemoryInterface: ClientLike + Sized {
    // Provided methods
    fn memory_doctor<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn memory_malloc_stats<R>(
        &self,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn memory_purge(&self) -> impl Future<Output = RedisResult<()>> + Send { ... }
    fn memory_stats<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn memory_usage<R, K>(
        &self,
        key: K,
        samples: Option<u32>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
}

Available on crate feature i-memory only.

Functions that implement the memory interface.

Provided Methods

fn memory_doctor<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

The MEMORY DOCTOR command reports about different memory-related issues that the Redis server experiences, and advises about possible remedies.

https://redis.io/commands/memory-doctor

fn memory_malloc_stats<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

The MEMORY MALLOC-STATS command provides an internal statistics report from the memory allocator.

https://redis.io/commands/memory-malloc-stats

fn memory_purge(&self) -> impl Future<Output = RedisResult<()>> + Send

The MEMORY PURGE command attempts to purge dirty pages so these can be reclaimed by the allocator.

https://redis.io/commands/memory-purge

fn memory_stats<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

The MEMORY STATS command returns an Array reply about the memory usage of the server.

https://redis.io/commands/memory-stats

fn memory_usage<R, K>( &self, key: K, samples: Option<u32>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

The MEMORY USAGE command reports the number of bytes that a key and its value require to be stored in RAM.

https://redis.io/commands/memory-usage

Object Safety

This trait is not object safe.