Trait fred::interfaces::PubsubInterface

pub trait PubsubInterface:
    ClientLike
    + Sized
    + Send {

    // Provided methods
    fn subscribe<S>(
        &self,
        channels: S,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<MultipleStrings> + Send { ... }
    fn unsubscribe<S>(
        &self,
        channels: S,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<MultipleStrings> + Send { ... }
    fn psubscribe<S>(
        &self,
        patterns: S,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<MultipleStrings> + Send { ... }
    fn punsubscribe<S>(
        &self,
        patterns: S,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where S: Into<MultipleStrings> + Send { ... }
    fn publish<R, S, V>(
        &self,
        channel: S,
        message: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn ssubscribe<C>(
        &self,
        channels: C,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where C: Into<MultipleStrings> + Send { ... }
    fn sunsubscribe<C>(
        &self,
        channels: C,
    ) -> impl Future<Output = RedisResult<()>> + Send
       where C: Into<MultipleStrings> + Send { ... }
    fn spublish<R, S, V>(
        &self,
        channel: S,
        message: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn pubsub_channels<R, S>(
        &self,
        pattern: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn pubsub_numpat<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn pubsub_numsub<R, S>(
        &self,
        channels: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<MultipleStrings> + Send { ... }
    fn pubsub_shardchannels<R, S>(
        &self,
        pattern: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send { ... }
    fn pubsub_shardnumsub<R, S>(
        &self,
        channels: S,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<MultipleStrings> + Send { ... }

}

Available on crate feature i-pubsub only.

Functions that implement the pubsub interface.

Provided Methods

fn subscribe<S>( &self, channels: S, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<MultipleStrings> + Send,

Subscribe to a channel on the publish-subscribe interface.

https://redis.io/commands/subscribe

fn unsubscribe<S>( &self, channels: S, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<MultipleStrings> + Send,

Unsubscribe from a channel on the PubSub interface.

https://redis.io/commands/unsubscribe

fn psubscribe<S>( &self, patterns: S, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<MultipleStrings> + Send,

Subscribes the client to the given patterns.

https://redis.io/commands/psubscribe

fn punsubscribe<S>( &self, patterns: S, ) -> impl Future<Output = RedisResult<()>> + Send
where S: Into<MultipleStrings> + Send,

Unsubscribes the client from the given patterns, or from all of them if none is given.

If no channels are provided this command returns an empty array.

https://redis.io/commands/punsubscribe

fn publish<R, S, V>( &self, channel: S, message: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Publish a message on the PubSub interface, returning the number of clients that received the message.

https://redis.io/commands/publish

fn ssubscribe<C>( &self, channels: C, ) -> impl Future<Output = RedisResult<()>> + Send
where C: Into<MultipleStrings> + Send,

Subscribes the client to the specified shard channels.

https://redis.io/commands/ssubscribe/

fn sunsubscribe<C>( &self, channels: C, ) -> impl Future<Output = RedisResult<()>> + Send
where C: Into<MultipleStrings> + Send,

Unsubscribes the client from the given shard channels, or from all of them if none is given.

If no channels are provided this command returns an empty array.

https://redis.io/commands/sunsubscribe/

fn spublish<R, S, V>( &self, channel: S, message: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Posts a message to the given shard channel.

https://redis.io/commands/spublish/

fn pubsub_channels<R, S>( &self, pattern: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

Lists the currently active channels.

https://redis.io/commands/pubsub-channels/

fn pubsub_numpat<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Returns the number of unique patterns that are subscribed to by clients.

https://redis.io/commands/pubsub-numpat/

fn pubsub_numsub<R, S>( &self, channels: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<MultipleStrings> + Send,

Returns the number of subscribers (exclusive of clients subscribed to patterns) for the specified channels.

https://redis.io/commands/pubsub-numsub/

fn pubsub_shardchannels<R, S>( &self, pattern: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send,

Lists the currently active shard channels.

https://redis.io/commands/pubsub-shardchannels/

fn pubsub_shardnumsub<R, S>( &self, channels: S, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<MultipleStrings> + Send,

Returns the number of subscribers for the specified shard channels.

https://redis.io/commands/pubsub-shardnumsub/

Object Safety

This trait is not object safe.