Trait fred::interfaces::SentinelInterface

pub trait SentinelInterface: ClientLike + Sized {

    // Provided methods
    fn ckquorum<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn flushconfig<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn failover<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn get_master_addr_by_name<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn info_cache<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn master<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn masters<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn monitor<R, N>(
        &self,
        name: N,
        ip: IpAddr,
        port: u16,
        quorum: u32,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn myid<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn pending_scripts<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn remove<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn replicas<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn sentinels<R, N>(
        &self,
        name: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send { ... }
    fn set<R, N, V>(
        &self,
        name: N,
        args: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             N: Into<Str> + Send,
             V: TryInto<RedisMap> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn simulate_failure<R>(
        &self,
        kind: SentinelFailureKind,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn reset<R, P>(
        &self,
        pattern: P,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             P: Into<Str> + Send { ... }
    fn config_get<R, K>(
        &self,
        name: K,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<Str> + Send { ... }
    fn config_set<R, K, V>(
        &self,
        name: K,
        value: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<Str> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }

}

Available on crate feature sentinel-client only.

Functions that implement the sentinel interface.

Provided Methods

fn ckquorum<R, N>(&self, name: N) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Check if the current Sentinel configuration is able to reach the quorum needed to failover a master, and the majority needed to authorize the failover.

fn flushconfig<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Force Sentinel to rewrite its configuration on disk, including the current Sentinel state.

fn failover<R, N>(&self, name: N) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Force a failover as if the master was not reachable, and without asking for agreement to other Sentinels.

fn get_master_addr_by_name<R, N>( &self, name: N, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Return the ip and port number of the master with that name.

fn info_cache<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Return cached INFO output from masters and replicas.

fn master<R, N>(&self, name: N) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Show the state and info of the specified master.

fn masters<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Show a list of monitored masters and their state.

fn monitor<R, N>( &self, name: N, ip: IpAddr, port: u16, quorum: u32, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Start Sentinel’s monitoring.

https://redis.io/topics/sentinel#reconfiguring-sentinel-at-runtime

fn myid<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

Return the ID of the Sentinel instance.

fn pending_scripts<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

This command returns information about pending scripts.

fn remove<R, N>(&self, name: N) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Stop Sentinel’s monitoring.

https://redis.io/topics/sentinel#reconfiguring-sentinel-at-runtime

fn replicas<R, N>(&self, name: N) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Show a list of replicas for this master, and their state.

fn sentinels<R, N>( &self, name: N, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send,

Show a list of sentinel instances for this master, and their state.

fn set<R, N, V>( &self, name: N, args: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, N: Into<Str> + Send, V: TryInto<RedisMap> + Send, V::Error: Into<RedisError> + Send,

Set Sentinel’s monitoring configuration.

https://redis.io/topics/sentinel#reconfiguring-sentinel-at-runtime

fn simulate_failure<R>( &self, kind: SentinelFailureKind, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

This command simulates different Sentinel crash scenarios.

fn reset<R, P>(&self, pattern: P) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, P: Into<Str> + Send,

This command will reset all the masters with matching name.

fn config_get<R, K>( &self, name: K, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<Str> + Send,

Get the current value of a global Sentinel configuration parameter. The specified name may be a wildcard, similar to the Redis CONFIG GET command.

fn config_set<R, K, V>( &self, name: K, value: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<Str> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Set the value of a global Sentinel configuration parameter.

Object Safety

This trait is not object safe.