Trait fred::interfaces::SlowlogInterface

pub trait SlowlogInterface: ClientLike + Sized {
    // Provided methods
    fn slowlog_get<R>(
        &self,
        count: Option<i64>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn slowlog_length<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis { ... }
    fn slowlog_reset(&self) -> impl Future<Output = RedisResult<()>> + Send { ... }
}

Available on crate feature i-slowlog only.

Functions that implement the slowlog interface.

Provided Methods

fn slowlog_get<R>( &self, count: Option<i64>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

This command is used to read the slow queries log.

https://redis.io/commands/slowlog#reading-the-slow-log

fn slowlog_length<R>(&self) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis,

This command is used to read length of the slow queries log.

https://redis.io/commands/slowlog#obtaining-the-current-length-of-the-slow-log

fn slowlog_reset(&self) -> impl Future<Output = RedisResult<()>> + Send

This command is used to reset the slow queries log.

https://redis.io/commands/slowlog#resetting-the-slow-log

Object Safety

This trait is not object safe.