Trait fred::interfaces::SortedSetsInterface

pub trait SortedSetsInterface: ClientLike + Sized {

    // Provided methods
    fn bzmpop<R, K>(
        &self,
        timeout: f64,
        keys: K,
        sort: ZCmp,
        count: Option<i64>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send { ... }
    fn bzpopmin<R, K>(
        &self,
        keys: K,
        timeout: f64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send { ... }
    fn bzpopmax<R, K>(
        &self,
        keys: K,
        timeout: f64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send { ... }
    fn zadd<R, K, V>(
        &self,
        key: K,
        options: Option<SetOptions>,
        ordering: Option<Ordering>,
        changed: bool,
        incr: bool,
        values: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<MultipleZaddValues> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn zcard<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zcount<R, K>(
        &self,
        key: K,
        min: f64,
        max: f64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zdiff<R, K>(
        &self,
        keys: K,
        withscores: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send { ... }
    fn zdiffstore<R, D, K>(
        &self,
        dest: D,
        keys: K,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             D: Into<RedisKey> + Send,
             K: Into<MultipleKeys> + Send { ... }
    fn zincrby<R, K, V>(
        &self,
        key: K,
        increment: f64,
        member: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn zinter<R, K, W>(
        &self,
        keys: K,
        weights: W,
        aggregate: Option<AggregateOptions>,
        withscores: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send,
             W: Into<MultipleWeights> + Send { ... }
    fn zinterstore<R, D, K, W>(
        &self,
        dest: D,
        keys: K,
        weights: W,
        aggregate: Option<AggregateOptions>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             D: Into<RedisKey> + Send,
             K: Into<MultipleKeys> + Send,
             W: Into<MultipleWeights> + Send { ... }
    fn zlexcount<R, K, M, N>(
        &self,
        key: K,
        min: M,
        max: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zpopmax<R, K>(
        &self,
        key: K,
        count: Option<usize>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zpopmin<R, K>(
        &self,
        key: K,
        count: Option<usize>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zmpop<R, K>(
        &self,
        keys: K,
        sort: ZCmp,
        count: Option<i64>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send { ... }
    fn zrandmember<R, K>(
        &self,
        key: K,
        count: Option<(i64, bool)>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zrangestore<R, D, S, M, N>(
        &self,
        dest: D,
        source: S,
        min: M,
        max: N,
        sort: Option<ZSort>,
        rev: bool,
        limit: Option<Limit>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             D: Into<RedisKey> + Send,
             S: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrange<R, K, M, N>(
        &self,
        key: K,
        min: M,
        max: N,
        sort: Option<ZSort>,
        rev: bool,
        limit: Option<Limit>,
        withscores: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrangebylex<R, K, M, N>(
        &self,
        key: K,
        min: M,
        max: N,
        limit: Option<Limit>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrevrangebylex<R, K, M, N>(
        &self,
        key: K,
        max: M,
        min: N,
        limit: Option<Limit>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrangebyscore<R, K, M, N>(
        &self,
        key: K,
        min: M,
        max: N,
        withscores: bool,
        limit: Option<Limit>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrevrangebyscore<R, K, M, N>(
        &self,
        key: K,
        max: M,
        min: N,
        withscores: bool,
        limit: Option<Limit>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrank<R, K, V>(
        &self,
        key: K,
        member: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn zrem<R, K, V>(
        &self,
        key: K,
        members: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<MultipleValues> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn zremrangebylex<R, K, M, N>(
        &self,
        key: K,
        min: M,
        max: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zremrangebyrank<R, K>(
        &self,
        key: K,
        start: i64,
        stop: i64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zremrangebyscore<R, K, M, N>(
        &self,
        key: K,
        min: M,
        max: N,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             M: TryInto<ZRange> + Send,
             M::Error: Into<RedisError> + Send,
             N: TryInto<ZRange> + Send,
             N::Error: Into<RedisError> + Send { ... }
    fn zrevrange<R, K>(
        &self,
        key: K,
        start: i64,
        stop: i64,
        withscores: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn zrevrank<R, K, V>(
        &self,
        key: K,
        member: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn zscore<R, K, V>(
        &self,
        key: K,
        member: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<RedisValue> + Send,
             V::Error: Into<RedisError> + Send { ... }
    fn zunion<R, K, W>(
        &self,
        keys: K,
        weights: W,
        aggregate: Option<AggregateOptions>,
        withscores: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<MultipleKeys> + Send,
             W: Into<MultipleWeights> + Send { ... }
    fn zunionstore<R, D, K, W>(
        &self,
        dest: D,
        keys: K,
        weights: W,
        aggregate: Option<AggregateOptions>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             D: Into<RedisKey> + Send,
             K: Into<MultipleKeys> + Send,
             W: Into<MultipleWeights> + Send { ... }
    fn zmscore<R, K, V>(
        &self,
        key: K,
        members: V,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             V: TryInto<MultipleValues> + Send,
             V::Error: Into<RedisError> + Send { ... }

}

Available on crate feature i-sorted-sets only.

Functions that implement the sorted sets interface.

Provided Methods

fn bzmpop<R, K>( &self, timeout: f64, keys: K, sort: ZCmp, count: Option<i64>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send,

The blocking variant of Self::zmpop.

https://redis.io/commands/bzmpop/

fn bzpopmin<R, K>( &self, keys: K, timeout: f64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send,

The blocking variant of Self::zpopmin.

https://redis.io/commands/bzpopmin

fn bzpopmax<R, K>( &self, keys: K, timeout: f64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send,

The blocking variant of Self::zpopmax.

https://redis.io/commands/bzpopmax

fn zadd<R, K, V>( &self, key: K, options: Option<SetOptions>, ordering: Option<Ordering>, changed: bool, incr: bool, values: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<MultipleZaddValues> + Send, V::Error: Into<RedisError> + Send,

Adds all the specified members with the specified scores to the sorted set stored at key.

https://redis.io/commands/zadd

fn zcard<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns the sorted set cardinality (number of elements) of the sorted set stored at key.

https://redis.io/commands/zcard

fn zcount<R, K>( &self, key: K, min: f64, max: f64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns the number of elements in the sorted set at key with a score between min and max.

https://redis.io/commands/zcount

fn zdiff<R, K>( &self, keys: K, withscores: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send,

This command is similar to ZDIFFSTORE, but instead of storing the resulting sorted set, it is returned to the client.

https://redis.io/commands/zdiff

fn zdiffstore<R, D, K>( &self, dest: D, keys: K, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, D: Into<RedisKey> + Send, K: Into<MultipleKeys> + Send,

Computes the difference between the first and all successive input sorted sets and stores the result in destination.

https://redis.io/commands/zdiffstore

fn zincrby<R, K, V>( &self, key: K, increment: f64, member: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Increments the score of member in the sorted set stored at key by increment.

https://redis.io/commands/zincrby

fn zinter<R, K, W>( &self, keys: K, weights: W, aggregate: Option<AggregateOptions>, withscores: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send, W: Into<MultipleWeights> + Send,

This command is similar to ZINTERSTORE, but instead of storing the resulting sorted set, it is returned to the client.

https://redis.io/commands/zinter

fn zinterstore<R, D, K, W>( &self, dest: D, keys: K, weights: W, aggregate: Option<AggregateOptions>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, D: Into<RedisKey> + Send, K: Into<MultipleKeys> + Send, W: Into<MultipleWeights> + Send,

Computes the intersection of the sorted sets given by the specified keys, and stores the result in destination.

https://redis.io/commands/zinterstore

fn zlexcount<R, K, M, N>( &self, key: K, min: M, max: N, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

When all the elements in a sorted set are inserted with the same score, in order to force lexicographical ordering, this command returns the number of elements in the sorted set at key with a value between min and max.

https://redis.io/commands/zlexcount

fn zpopmax<R, K>( &self, key: K, count: Option<usize>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Removes and returns up to count members with the highest scores in the sorted set stored at key.

https://redis.io/commands/zpopmax

fn zpopmin<R, K>( &self, key: K, count: Option<usize>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Removes and returns up to count members with the lowest scores in the sorted set stored at key.

https://redis.io/commands/zpopmin

fn zmpop<R, K>( &self, keys: K, sort: ZCmp, count: Option<i64>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send,

Pops one or more elements, that are member-score pairs, from the first non-empty sorted set in the provided list of key names.

https://redis.io/commands/zmpop/

fn zrandmember<R, K>( &self, key: K, count: Option<(i64, bool)>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

When called with just the key argument, return a random element from the sorted set value stored at key.

https://redis.io/commands/zrandmember

fn zrangestore<R, D, S, M, N>( &self, dest: D, source: S, min: M, max: N, sort: Option<ZSort>, rev: bool, limit: Option<Limit>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, D: Into<RedisKey> + Send, S: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

This command is like ZRANGE, but stores the result in the destination key.

https://redis.io/commands/zrangestore

fn zrange<R, K, M, N>( &self, key: K, min: M, max: N, sort: Option<ZSort>, rev: bool, limit: Option<Limit>, withscores: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

Returns the specified range of elements in the sorted set stored at key.

https://redis.io/commands/zrange

fn zrangebylex<R, K, M, N>( &self, key: K, min: M, max: N, limit: Option<Limit>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

When all the elements in a sorted set are inserted with the same score, in order to force lexicographical ordering, this command returns all the elements in the sorted set at key with a value between min and max.

https://redis.io/commands/zrangebylex

fn zrevrangebylex<R, K, M, N>( &self, key: K, max: M, min: N, limit: Option<Limit>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

When all the elements in a sorted set are inserted with the same score, in order to force lexicographical ordering, this command returns all the elements in the sorted set at key with a value between max and min.

https://redis.io/commands/zrevrangebylex

fn zrangebyscore<R, K, M, N>( &self, key: K, min: M, max: N, withscores: bool, limit: Option<Limit>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

Returns all the elements in the sorted set at key with a score between min and max (including elements with score equal to min or max).

https://redis.io/commands/zrangebyscore

fn zrevrangebyscore<R, K, M, N>( &self, key: K, max: M, min: N, withscores: bool, limit: Option<Limit>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

Returns all the elements in the sorted set at key with a score between max and min (including elements with score equal to max or min).

https://redis.io/commands/zrevrangebyscore

fn zrank<R, K, V>( &self, key: K, member: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Returns the rank of member in the sorted set stored at key, with the scores ordered from low to high.

https://redis.io/commands/zrank

fn zrem<R, K, V>( &self, key: K, members: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<MultipleValues> + Send, V::Error: Into<RedisError> + Send,

Removes the specified members from the sorted set stored at key. Non existing members are ignored.

https://redis.io/commands/zrem

fn zremrangebylex<R, K, M, N>( &self, key: K, min: M, max: N, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

When all the elements in a sorted set are inserted with the same score, in order to force lexicographical ordering, this command removes all elements in the sorted set stored at key between the lexicographical range specified by min and max.

https://redis.io/commands/zremrangebylex

fn zremrangebyrank<R, K>( &self, key: K, start: i64, stop: i64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Removes all elements in the sorted set stored at key with rank between start and stop.

https://redis.io/commands/zremrangebyrank

fn zremrangebyscore<R, K, M, N>( &self, key: K, min: M, max: N, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, M: TryInto<ZRange> + Send, M::Error: Into<RedisError> + Send, N: TryInto<ZRange> + Send, N::Error: Into<RedisError> + Send,

Removes all elements in the sorted set stored at key with a score between min and max.

https://redis.io/commands/zremrangebyscore

fn zrevrange<R, K>( &self, key: K, start: i64, stop: i64, withscores: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Returns the specified range of elements in the sorted set stored at key.

https://redis.io/commands/zrevrange

fn zrevrank<R, K, V>( &self, key: K, member: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Returns the rank of member in the sorted set stored at key, with the scores ordered from high to low.

https://redis.io/commands/zrevrank

fn zscore<R, K, V>( &self, key: K, member: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<RedisValue> + Send, V::Error: Into<RedisError> + Send,

Returns the score of member in the sorted set at key.

https://redis.io/commands/zscore

fn zunion<R, K, W>( &self, keys: K, weights: W, aggregate: Option<AggregateOptions>, withscores: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<MultipleKeys> + Send, W: Into<MultipleWeights> + Send,

This command is similar to ZUNIONSTORE, but instead of storing the resulting sorted set, it is returned to the client.

https://redis.io/commands/zunion

fn zunionstore<R, D, K, W>( &self, dest: D, keys: K, weights: W, aggregate: Option<AggregateOptions>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, D: Into<RedisKey> + Send, K: Into<MultipleKeys> + Send, W: Into<MultipleWeights> + Send,

Computes the union of the sorted sets given by the specified keys, and stores the result in destination.

https://redis.io/commands/zunionstore

fn zmscore<R, K, V>( &self, key: K, members: V, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, V: TryInto<MultipleValues> + Send, V::Error: Into<RedisError> + Send,

Returns the scores associated with the specified members in the sorted set stored at key.

https://redis.io/commands/zmscore

Object Safety

This trait is not object safe.