Trait fred::interfaces::TimeSeriesInterface

pub trait TimeSeriesInterface: ClientLike {

    // Provided methods
    fn ts_add<R, K, T, L>(
        &self,
        key: K,
        timestamp: T,
        value: f64,
        retention: Option<u64>,
        encoding: Option<Encoding>,
        chunk_size: Option<u64>,
        on_duplicate: Option<DuplicatePolicy>,
        labels: L,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             T: TryInto<Timestamp> + Send,
             T::Error: Into<RedisError> + Send,
             L: TryInto<RedisMap> + Send,
             L::Error: Into<RedisError> { ... }
    fn ts_alter<R, K, L>(
        &self,
        key: K,
        retention: Option<u64>,
        chunk_size: Option<u64>,
        duplicate_policy: Option<DuplicatePolicy>,
        labels: L,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             L: TryInto<RedisMap> + Send,
             L::Error: Into<RedisError> { ... }
    fn ts_create<R, K, L>(
        &self,
        key: K,
        retention: Option<u64>,
        encoding: Option<Encoding>,
        chunk_size: Option<u64>,
        duplicate_policy: Option<DuplicatePolicy>,
        labels: L,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             L: TryInto<RedisMap> + Send,
             L::Error: Into<RedisError> { ... }
    fn ts_createrule<R, S, D>(
        &self,
        src: S,
        dest: D,
        aggregation: (Aggregator, u64),
        align_timestamp: Option<u64>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<RedisKey> + Send,
             D: Into<RedisKey> + Send { ... }
    fn ts_decrby<R, K, L>(
        &self,
        key: K,
        subtrahend: f64,
        timestamp: Option<Timestamp>,
        retention: Option<u64>,
        uncompressed: bool,
        chunk_size: Option<u64>,
        labels: L,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             L: TryInto<RedisMap> + Send,
             L::Error: Into<RedisError> + Send { ... }
    fn ts_del<R, K>(
        &self,
        key: K,
        from: i64,
        to: i64,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn ts_deleterule<R, S, D>(
        &self,
        src: S,
        dest: D,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<RedisKey> + Send,
             D: Into<RedisKey> + Send { ... }
    fn ts_get<R, K>(
        &self,
        key: K,
        latest: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn ts_incrby<R, K, L>(
        &self,
        key: K,
        addend: f64,
        timestamp: Option<Timestamp>,
        retention: Option<u64>,
        uncompressed: bool,
        chunk_size: Option<u64>,
        labels: L,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             L: TryInto<RedisMap> + Send,
             L::Error: Into<RedisError> + Send { ... }
    fn ts_info<R, K>(
        &self,
        key: K,
        debug: bool,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send { ... }
    fn ts_madd<R, K, I>(
        &self,
        samples: I,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             I: IntoIterator<Item = (K, Timestamp, f64)> + Send { ... }
    fn ts_mget<R, L, S, I>(
        &self,
        latest: bool,
        labels: Option<L>,
        filters: I,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             L: Into<GetLabels> + Send,
             S: Into<Str> + Send,
             I: IntoIterator<Item = S> + Send { ... }
    fn ts_mrange<R, F, T, I, S, J>(
        &self,
        from: F,
        to: T,
        latest: bool,
        filter_by_ts: I,
        filter_by_value: Option<(i64, i64)>,
        labels: Option<GetLabels>,
        count: Option<u64>,
        aggregation: Option<RangeAggregation>,
        filters: J,
        group_by: Option<GroupBy>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             F: TryInto<GetTimestamp> + Send,
             F::Error: Into<RedisError> + Send,
             T: TryInto<GetTimestamp> + Send,
             T::Error: Into<RedisError> + Send,
             S: Into<Str> + Send,
             I: IntoIterator<Item = i64> + Send,
             J: IntoIterator<Item = S> + Send { ... }
    fn ts_mrevrange<R, F, T, I, S, J>(
        &self,
        from: F,
        to: T,
        latest: bool,
        filter_by_ts: I,
        filter_by_value: Option<(i64, i64)>,
        labels: Option<GetLabels>,
        count: Option<u64>,
        aggregation: Option<RangeAggregation>,
        filters: J,
        group_by: Option<GroupBy>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             F: TryInto<GetTimestamp> + Send,
             F::Error: Into<RedisError> + Send,
             T: TryInto<GetTimestamp> + Send,
             T::Error: Into<RedisError> + Send,
             S: Into<Str> + Send,
             I: IntoIterator<Item = i64> + Send,
             J: IntoIterator<Item = S> + Send { ... }
    fn ts_queryindex<R, S, I>(
        &self,
        filters: I,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             S: Into<Str> + Send,
             I: IntoIterator<Item = S> + Send { ... }
    fn ts_range<R, K, F, T, I>(
        &self,
        key: K,
        from: F,
        to: T,
        latest: bool,
        filter_by_ts: I,
        filter_by_value: Option<(i64, i64)>,
        count: Option<u64>,
        aggregation: Option<RangeAggregation>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: TryInto<GetTimestamp> + Send,
             F::Error: Into<RedisError> + Send,
             T: TryInto<GetTimestamp> + Send,
             T::Error: Into<RedisError> + Send,
             I: IntoIterator<Item = i64> + Send { ... }
    fn ts_revrange<R, K, F, T, I>(
        &self,
        key: K,
        from: F,
        to: T,
        latest: bool,
        filter_by_ts: I,
        filter_by_value: Option<(i64, i64)>,
        count: Option<u64>,
        aggregation: Option<RangeAggregation>,
    ) -> impl Future<Output = RedisResult<R>> + Send
       where R: FromRedis,
             K: Into<RedisKey> + Send,
             F: TryInto<GetTimestamp> + Send,
             F::Error: Into<RedisError> + Send,
             T: TryInto<GetTimestamp> + Send,
             T::Error: Into<RedisError> + Send,
             I: IntoIterator<Item = i64> + Send { ... }

}

Available on crate feature i-time-series only.

A Redis Timeseries interface.

Provided Methods

fn ts_add<R, K, T, L>( &self, key: K, timestamp: T, value: f64, retention: Option<u64>, encoding: Option<Encoding>, chunk_size: Option<u64>, on_duplicate: Option<DuplicatePolicy>, labels: L, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, T: TryInto<Timestamp> + Send, T::Error: Into<RedisError> + Send, L: TryInto<RedisMap> + Send, L::Error: Into<RedisError>,

Append a sample to a time series.

https://redis.io/commands/ts.add/

fn ts_alter<R, K, L>( &self, key: K, retention: Option<u64>, chunk_size: Option<u64>, duplicate_policy: Option<DuplicatePolicy>, labels: L, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, L: TryInto<RedisMap> + Send, L::Error: Into<RedisError>,

Update the retention, chunk size, duplicate policy, and labels of an existing time series.

https://redis.io/commands/ts.alter/

fn ts_create<R, K, L>( &self, key: K, retention: Option<u64>, encoding: Option<Encoding>, chunk_size: Option<u64>, duplicate_policy: Option<DuplicatePolicy>, labels: L, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, L: TryInto<RedisMap> + Send, L::Error: Into<RedisError>,

Create a new time series.

https://redis.io/commands/ts.create/

fn ts_createrule<R, S, D>( &self, src: S, dest: D, aggregation: (Aggregator, u64), align_timestamp: Option<u64>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<RedisKey> + Send, D: Into<RedisKey> + Send,

Create a compaction rule.

https://redis.io/commands/ts.createrule/

fn ts_decrby<R, K, L>( &self, key: K, subtrahend: f64, timestamp: Option<Timestamp>, retention: Option<u64>, uncompressed: bool, chunk_size: Option<u64>, labels: L, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, L: TryInto<RedisMap> + Send, L::Error: Into<RedisError> + Send,

Decrease the value of the sample with the maximum existing timestamp, or create a new sample with a value equal to the value of the sample with the maximum existing timestamp with a given decrement.

https://redis.io/commands/ts.decrby/

fn ts_del<R, K>( &self, key: K, from: i64, to: i64, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Delete all samples between two timestamps for a given time series.

https://redis.io/commands/ts.del/

fn ts_deleterule<R, S, D>( &self, src: S, dest: D, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<RedisKey> + Send, D: Into<RedisKey> + Send,

Delete a compaction rule.

https://redis.io/commands/ts.deleterule/

fn ts_get<R, K>( &self, key: K, latest: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Get the sample with the highest timestamp from a given time series.

https://redis.io/commands/ts.get/

fn ts_incrby<R, K, L>( &self, key: K, addend: f64, timestamp: Option<Timestamp>, retention: Option<u64>, uncompressed: bool, chunk_size: Option<u64>, labels: L, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, L: TryInto<RedisMap> + Send, L::Error: Into<RedisError> + Send,

Increase the value of the sample with the maximum existing timestamp, or create a new sample with a value equal to the value of the sample with the maximum existing timestamp with a given increment.

https://redis.io/commands/ts.incrby/

fn ts_info<R, K>( &self, key: K, debug: bool, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send,

Return information and statistics for a time series.

https://redis.io/commands/ts.info/

fn ts_madd<R, K, I>( &self, samples: I, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, I: IntoIterator<Item = (K, Timestamp, f64)> + Send,

Append new samples to one or more time series.

https://redis.io/commands/ts.madd/

fn ts_mget<R, L, S, I>( &self, latest: bool, labels: Option<L>, filters: I, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, L: Into<GetLabels> + Send, S: Into<Str> + Send, I: IntoIterator<Item = S> + Send,

Get the sample with the highest timestamp from each time series matching a specific filter.

See Resp2TimeSeriesValues and Resp3TimeSeriesValues for more information.

https://redis.io/commands/ts.mget/

fn ts_mrange<R, F, T, I, S, J>( &self, from: F, to: T, latest: bool, filter_by_ts: I, filter_by_value: Option<(i64, i64)>, labels: Option<GetLabels>, count: Option<u64>, aggregation: Option<RangeAggregation>, filters: J, group_by: Option<GroupBy>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, F: TryInto<GetTimestamp> + Send, F::Error: Into<RedisError> + Send, T: TryInto<GetTimestamp> + Send, T::Error: Into<RedisError> + Send, S: Into<Str> + Send, I: IntoIterator<Item = i64> + Send, J: IntoIterator<Item = S> + Send,

Query a range across multiple time series by filters in the forward direction.

See Resp2TimeSeriesValues and Resp3TimeSeriesValues for more information.

https://redis.io/commands/ts.mrange/

fn ts_mrevrange<R, F, T, I, S, J>( &self, from: F, to: T, latest: bool, filter_by_ts: I, filter_by_value: Option<(i64, i64)>, labels: Option<GetLabels>, count: Option<u64>, aggregation: Option<RangeAggregation>, filters: J, group_by: Option<GroupBy>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, F: TryInto<GetTimestamp> + Send, F::Error: Into<RedisError> + Send, T: TryInto<GetTimestamp> + Send, T::Error: Into<RedisError> + Send, S: Into<Str> + Send, I: IntoIterator<Item = i64> + Send, J: IntoIterator<Item = S> + Send,

Query a range across multiple time series by filters in the reverse direction.

See Resp2TimeSeriesValues and Resp3TimeSeriesValues for more information.

https://redis.io/commands/ts.mrevrange/

fn ts_queryindex<R, S, I>( &self, filters: I, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, S: Into<Str> + Send, I: IntoIterator<Item = S> + Send,

Get all time series keys matching a filter list.

https://redis.io/commands/ts.queryindex/

fn ts_range<R, K, F, T, I>( &self, key: K, from: F, to: T, latest: bool, filter_by_ts: I, filter_by_value: Option<(i64, i64)>, count: Option<u64>, aggregation: Option<RangeAggregation>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: TryInto<GetTimestamp> + Send, F::Error: Into<RedisError> + Send, T: TryInto<GetTimestamp> + Send, T::Error: Into<RedisError> + Send, I: IntoIterator<Item = i64> + Send,

Query a range in forward direction.

https://redis.io/commands/ts.range/

fn ts_revrange<R, K, F, T, I>( &self, key: K, from: F, to: T, latest: bool, filter_by_ts: I, filter_by_value: Option<(i64, i64)>, count: Option<u64>, aggregation: Option<RangeAggregation>, ) -> impl Future<Output = RedisResult<R>> + Send
where R: FromRedis, K: Into<RedisKey> + Send, F: TryInto<GetTimestamp> + Send, F::Error: Into<RedisError> + Send, T: TryInto<GetTimestamp> + Send, T::Error: Into<RedisError> + Send, I: IntoIterator<Item = i64> + Send,

Query a range in reverse direction.

https://redis.io/commands/ts.revrange/

Object Safety

This trait is not object safe.