Module fred::sync::borrowed

A remote interface for a RedisClient that borrows self on each command. This is ideal for use within a wrapping struct that manages ownership over inner values.

Structs

RedisClientRemote
A Send and Sync wrapper around a redis client that borrows self on each command, instead of taking ownership over self. This pattern gives the caller more freedom to manage ownership over the value, such as when the client is wrapped with another struct.