Struct fred::sync::owned::RedisClientRemote

pub struct RedisClientRemote { /* fields omitted */ }

A Send and Sync wrapper around a RedisClient.

This module exposes the same interface as the RedisClient struct, but can be safely sent and used across threads.

The underlying implementation uses message passing patterns to communicate with the RedisClient instance on the event loop thread. As a result all commands, with the exception of state(), have some added overhead due to the inter-thread communication logic. The only synchronized code path on each command is reading the client's state, which is wrapped in a RwLock.

See examples/sync_owned.rs for usage examples.

Methods

impl RedisClientRemote

fn new() -> RedisClientRemote

Create a new, empty RedisClientRemote.

fn from_borrowed(client: RedisClientRemote) -> RedisClientRemote

Create from a borrowed instance.

fn to_borrowed(&self) -> Option<RedisClientRemote>

Attempt to convert to a borrowed instance. This returns None if init has not yet been called.

fn into_borrowed(self) -> Option<RedisClientRemote>

Attempt to convert this owned instance into a borrowed instance.

fn inner_borrowed(&self) -> &Arc<RwLock<Option<RedisClientBorrowed>>>

fn init(
&self,
client: RedisClient
) -> Box<Future<Item = RedisClient, Error = RedisError>>

Initialize the remote interface.

This function must run on the same thread that created the RedisClient.

fn on_connect(&self) -> Box<Future<Item = (), Error = RedisError>>

Returns a future that resolves when the underlying client connects to the server. This function can act as a convenient way of notifying a separate thread when the client has connected to the server and can begin processing commands.

This can be called before init if needed, however the callback will not be registered on the underlying client until init is called. For this reason it's best to call init with a client that has not yet starting running its connection future.

See the examples/sync_borrowed.rs file for usage examples.

fn subscribe<K: Into<String>>(
self,
channel: K
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Subscribe to a channel on the PubSub interface. Any messages received before on_message is called will be discarded, so it's usually best to call on_message before calling subscribe for the first time. The usize returned here is the number of channels to which the client is currently subscribed.

https://redis.io/commands/subscribe

fn unsubscribe<K: Into<String>>(
self,
channel: K
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Unsubscribe from a channel on the PubSub interface.

https://redis.io/commands/unsubscribe

fn publish<K: Into<String>, V: Into<RedisValue>>(
self,
channel: K,
message: V
) -> Box<Future<Item = (Self, i64), Error = RedisError>>

Publish a message on the PubSub interface, returning the number of clients that received the message.

https://redis.io/commands/publish

fn get<K: Into<RedisKey>>(
self,
key: K
) -> Box<Future<Item = (Self, Option<RedisValue>), Error = RedisError>>

Read a value from Redis at key.

https://redis.io/commands/get

fn set<K: Into<RedisKey>, V: Into<RedisValue>>(
self,
key: K,
value: V,
expire: Option<Expiration>,
options: Option<SetOptions>
) -> Box<Future<Item = (Self, bool), Error = RedisError>>

Set a value at key with optional NX|XX and EX|PX arguments. The bool returned by this function describes whether or not the key was set due to any NX|XX options.

https://redis.io/commands/set

fn del<K: Into<MultipleKeys>>(
self,
keys: K
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Removes the specified keys. A key is ignored if it does not exist. Returns the number of keys removed.

https://redis.io/commands/del

fn decr<K: Into<RedisKey>>(
self,
key: K
) -> Box<Future<Item = (Self, i64), Error = RedisError>>

Decrements the number stored at key by one. If the key does not exist, it is set to 0 before performing the operation. Returns error if the key contains a value of the wrong type.

https://redis.io/commands/decr

fn incr<K: Into<RedisKey>>(
self,
key: K
) -> Box<Future<Item = (Self, i64), Error = RedisError>>

Increments the number stored at key by one. If the key does not exist, it is set to 0 before performing the operation. Returns error if the value at key is of wrong type.

https://redis.io/commands/incr

fn incrby<K: Into<RedisKey>>(
self,
key: K,
incr: i64
) -> Box<Future<Item = (Self, i64), Error = RedisError>>

Increments the number stored at key by incr. If the key does not exist, it is set to 0 before performing the operation. Returns an error if the value at key is of the wrong type.

https://redis.io/commands/incrby

fn incrbyfloat<K: Into<RedisKey>>(
self,
key: K,
incr: f64
) -> Box<Future<Item = (Self, f64), Error = RedisError>>

Increment the string representing a floating point number stored at key by the argument value. If the key does not exist, it is set to 0 before performing the operation. Returns error if key value is wrong type or if the current value or increment value are not parseable as float value.

https://redis.io/commands/incrbyfloat

fn hget<F: Into<RedisKey>, K: Into<RedisKey>>(
self,
key: K,
field: F
) -> Box<Future<Item = (Self, Option<RedisValue>), Error = RedisError>>

Returns the value associated with field in the hash stored at key.

https://redis.io/commands/hget

fn hset<K: Into<RedisKey>, F: Into<RedisKey>, V: Into<RedisValue>>(
self,
key: K,
field: F,
value: V
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Sets field in the hash stored at key to value. If key does not exist, a new key holding a hash is created. If field already exists in the hash, it is overwritten. Note: Return value of 1 means new field was created and set. Return of 0 means field already exists and was overwritten.

https://redis.io/commands/hset

fn hdel<K: Into<RedisKey>, F: Into<MultipleKeys>>(
self,
key: K,
fields: F
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Removes the specified fields from the hash stored at key. Specified fields that do not exist within this hash are ignored. If key does not exist, it is treated as an empty hash and this command returns 0.

https://redis.io/commands/hdel

fn hlen<K: Into<RedisKey>>(
self,
key: K
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Returns the number of fields contained in the hash stored at key.

https://redis.io/commands/hlen

fn hmget<F: Into<MultipleKeys>, K: Into<RedisKey>>(
self,
key: K,
fields: F
) -> Box<Future<Item = (Self, Vec<RedisValue>), Error = RedisError>>

Returns the values associated with the specified fields in the hash stored at key. Values in a returned list may be null.

https://redis.io/commands/hmget

fn hmset<V: Into<RedisValue>, F: Into<RedisKey> + Hash + Eq, K: Into<RedisKey>>(
self,
key: K,
values: HashMap<F, V>
) -> Box<Future<Item = (Self, String), Error = RedisError>>

Sets the specified fields to their respective values in the hash stored at key. This command overwrites any specified fields already existing in the hash. If key does not exist, a new key holding a hash is created.

https://redis.io/commands/hmset

fn hsetnx<K: Into<RedisKey>, F: Into<RedisKey>, V: Into<RedisValue>>(
self,
key: K,
field: F,
value: V
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Sets field in the hash stored at key to value, only if field does not yet exist. If key does not exist, a new key holding a hash is created. Note: Return value of 1 means new field was created and set. Return of 0 means no operation performed.

https://redis.io/commands/hsetnx

fn hstrlen<K: Into<RedisKey>, F: Into<RedisKey>>(
self,
key: K,
field: F
) -> Box<Future<Item = (Self, usize), Error = RedisError>>

Returns the string length of the value associated with field in the hash stored at key. If the key or the field do not exist, 0 is returned.

https://redis.io/commands/hstrlen

fn hvals<K: Into<RedisKey>>(
self,
key: K
) -> Box<Future<Item = (Self, Vec<RedisValue>), Error = RedisError>>

Returns all values in the hash stored at key. Returns an empty vector if the list is empty.

https://redis.io/commands/hvals

fn hkeys<K: Into<RedisKey>>(
self,
key: K
) -> Box<Future<Item = (Self, Vec<String>), Error = RedisError>>

Returns all field names in the hash stored at key. Returns an empty vec if the list is empty. Null fields are converted to "nil".

https://redis.io/commands/hkeys