Module fred::types

The structs and enums used by the Redis client.

Structs

BackpressureConfig

Configuration options for backpressure features in the client.

Builder

A client and pool builder interface.

ClusterInfoi-cluster

A parsed response from the CLUSTER INFO command.

ClusterRouting

The cached view of the cluster used by the client to route commands to the correct cluster nodes.

ConnectionConfig

Configuration options related to the creation or management of TCP connection.

CustomCommand

Configuration for custom redis commands, primarily used for interacting with third party modules or extensions.

DatabaseMemoryStatsi-memory

The parsed result of the MEMORY STATS command for a specific database.

FtAggregateOptionsi-redisearch

Arguments to the FT.AGGREGATE command.

FtAlterOptionsi-redisearch

Arguments to FT.ALTER.

FtCreateOptionsi-redisearch

Arguments for FT.CREATE.

FtSearchOptionsi-redisearch

Arguments to FT.SEARCH.

Functioni-scripts

An individual function within a Library.

GeoPositioni-geo

A struct describing the longitude and latitude coordinates of a GEO command.

GeoRadiusInfoi-geo

A typed struct representing the full output of the GEORADIUS (or similar) command.

GeoValuei-geo

A struct describing the value inside a GEO data structure.

GroupByi-time-series

A struct representing GROUPBY label REDUCE reducer in commands such as TS.MRANGE.

HScanResult

The result of a HSCAN operation.

Invalidationi-client and i-tracking

A client tracking invalidation message from the provided server.

KeyspaceEvent

An event on the publish-subscribe interface describing a keyspace notification.

Libraryi-scripts

A helper struct for interacting with libraries and functions.

MemoryStatsi-memory

The parsed result of the MEMORY STATS command.

Message

A publish-subscribe message.

MultipleGeoValuesi-geo

A convenience struct for commands that take one or more GEO values.

MultipleHashSlots

A convenience struct for functions that take one or more hash slot values.

MultipleIDsi-streams

One or more IDs for elements in a stream.

MultipleKeys

Convenience struct for commands that take 1 or more keys.

MultipleOrderedPairsi-streams

One or more ordered key-value pairs, typically used as an argument for XADD.

MultipleWeightsi-sorted-sets

Convenience struct for ZINTERSTORE and ZUNIONSTORE when accepting 1 or more weights arguments.

MultipleZaddValuesi-sorted-sets

Convenience struct for the ZADD command to accept 1 or more (score, value) arguments.

Options

Options to configure or overwrite for individual commands.

PerformanceConfig

Configuration options that can affect the performance of the client.

RangeAggregationi-time-series

A struct representing [ALIGN align] AGGREGATION aggregator bucketDuration [BUCKETTIMESTAMP bt] [EMPTY] in commands such as TS.MRANGE.

RedisConfig

Configuration options for a RedisClient.

RedisKey

A key in Redis.

RedisMap

A map of (RedisKey, RedisValue) pairs.

ReplicaConfigreplicas

Configuration options for replica node connections.

SScanResult

The result of a SSCAN operation.

ScanResult

The result of a SCAN operation.

Scripti-scripts

An interface for caching and running lua scripts.

SearchFieldi-redisearch

A search field with an optional property.

SearchFilteri-redisearch

Arguments for FILTER in FT.SEARCH.

SearchGeoFilteri-redisearch

Arguments for GEOFILTER in FT.SEARCH.

SearchHighlighti-redisearch

Arguments used in HIGHLIGHT values.

SearchParameteri-redisearch

Arguments for PARAMS in FT.AGGREGATE.

SearchReduceri-redisearch

REDUCE arguments in FT.AGGREGATE.

SearchSchemai-redisearch

Arguments for SCHEMA in FT.CREATE.

SearchSortByi-redisearch

Arguments for SORTBY in FT.SEARCH.

SearchSummarizei-redisearch

Arguments used in SUMMARIZE values.

SentinelConfigsentinel-client

Configuration options for sentinel clients.

Server

State necessary to identify or connect to a server.

SlotRange

A slot range and associated cluster node information from the CLUSTER SLOTS command.

SlowlogEntry

The output of an entry in the slow queries log.

Statsmetrics

Stats describing a distribution of samples.

TcpConfig

TCP configuration options.

TlsConfigenable-rustls or enable-native-tls or enable-rustls-ring

TLS configuration for a client.

TracingConfigpartial-tracing

Configuration options for tracing.

UnresponsiveConfig

Configuration options used to detect potentially unresponsive connections.

Version

SemVer version as defined by https://semver.org.

WithCursori-redisearch

Arguments for WITHCURSOR in FT.AGGREGATE.

XCapi-streams

Stream cap arguments for XADD, XTRIM, etc.

XPendingArgsi-streams

A struct representing the trailing optional arguments to XPENDING.

ZRangei-sorted-sets

A wrapper struct for a range bound in a sorted set command.

ZScanResult

The result of a ZSCAN operation.

Enums

AggregateOperationi-redisearch

An aggregation operation used in FT.AGGREGATE.

AggregateOptions

Aggregate options for the zinterstore (and related) commands.

Aggregatori-time-series

An aggregation policy to use with certain timeseries commands.

BackpressurePolicy

Backpressure policies to apply when the max number of in-flight commands is reached on a connection.

Blocking

Describes how the client should respond when a command is sent while the client is in a blocked state from a blocking command.

BucketTimestampi-time-series

A BUCKETTIMESTAMP argument in commands such as TS.MRANGE.

ClientKillFilteri-client

Filters provided to the CLIENT KILL command.

ClientKillTypei-client

The type of clients to close.

ClientPauseKindi-client

Filters for the CLIENT PAUSE command.

ClientReplyFlagi-client

Arguments for the CLIENT REPLY command.

ClientState

The state of the underlying connection to the Redis server.

ClientUnblockFlag

Arguments to the CLIENT UNBLOCK command.

ClusterDiscoveryPolicy

A policy that determines how clustered clients initially connect to and discover other cluster nodes.

ClusterFailoverFlagi-cluster

Options for the CLUSTER FAILOVER command.

ClusterHash

A cluster hashing policy.

ClusterResetFlagi-cluster

Flags for the CLUSTER RESET command.

ClusterSetSlotStatei-cluster

Flags for the CLUSTER SETSLOT command.

ClusterStatei-cluster

The state of the cluster from the CLUSTER INFO command.

ClusterStateChange

An enum describing the possible ways in which a Redis cluster can change state.

DuplicatePolicyi-time-series

The duplicate policy used with certain timeseries commands.

Encodingi-time-series

Encoding arguments for certain timeseries commands.

Expiration

Expiration options for the set command.

ExpireOptions

Options for certain expiration commands (PEXPIRE, etc).

FnPolicy

The policy type for the FUNCTION RESTORE command.

FunctionFlagi-scripts

Possible flags associated with a Function.

GeoUniti-geo

Units for the GEO DIST command.

GetLabelsi-time-series

Arguments equivalent to WITHLABELS | SELECTED_LABELS label... in various time series GET functions.

GetTimestampi-time-series

A timestamp query used in commands such as TS.MRANGE.

IndexKindi-redisearch

Index arguments for FT.CREATE.

InfoKind

Options for the info command.

LMoveDirectioni-lists

The direction to move elements in a *LMOVE command.

ListLocationi-lists

Location flag for the LINSERT command.

Loadi-redisearch

Arguments to LOAD in FT.AGGREGATE.

MessageKind

The kind of pubsub message.

Orderingi-sorted-sets

Ordering options for the ZADD (and related) commands.

ReconnectErrorcustom-reconnect-errors

Special errors that can trigger reconnection logic, which can also retry the failing command if possible.

ReconnectPolicy

The type of reconnection policy to use. This will apply to every connection used by the client.

RedisValue

A value used in a Redis command.

RedisValueKind

The kind of value from Redis.

Reduceri-time-series

A REDUCER argument in commands such as TS.MRANGE.

ReducerFunci-redisearch

GROUPBY reducer functions.

Resp3Frame

A RESP3 frame that uses Bytes and Str as the underlying buffer type.

RespVersion

The RESP version used in the HELLO request.

ScanType

The types of values supported by the type command.

ScriptDebugFlag

Flags for the SCRIPT DEBUG command.

SearchSchemaKindi-redisearch

One of the available schema types used with FT.CREATE or FT.ALTER.

SentinelFailureKindsentinel-client

Arguments for the SENTINEL SIMULATE-FAILURE command.

ServerConfig

Connection configuration for the Redis server.

SetOptions

Options for the set command.

ShutdownFlags

Arguments passed to the SHUTDOWN command.

SortOrder

The sort order for redis commands that take or return a sorted list.

SpellcheckTermsi-redisearch

Arguments to TERMS in FT.SPELLCHECK,

StringOrNumber

An argument representing a string or number.

Timestampi-time-series

A timestamp used in most timeseries commands.

TlsConnectorenable-rustls or enable-native-tls or enable-rustls-ring

An enum for interacting with various TLS libraries and interfaces.

TlsHostMappingenable-rustls or enable-native-tls or enable-rustls-ring

An optional enum used to describe how the client should modify or map IP addresses and hostnames in a clustered deployment.

Togglei-client and i-tracking

An ON|OFF flag used with client tracking commands.

XCapKindi-streams

The MAXLEN or MINID argument for a stream cap.

XCapTrimi-streams

Representation for the “=” or “~” operator in XADD, etc.

XIDi-streams

Stream ID arguments for XADD, XREAD, etc.

ZCmpi-sorted-sets

MIN|MAX arguments for BZMPOP, etc.

ZRangeBoundi-sorted-sets

An index, score, lexicographical, or +|-|+inf|-inf range bound for the ZRANGE command.

ZRangeKindi-sorted-sets

The type of range interval bound.

ZSorti-sorted-sets

Options for the ZRANGE (and related) commands.

Constants

DEFAULT_JITTER_MS

The default amount of jitter when waiting to reconnect.

Traits

FromRedis

A trait used to convert various forms of RedisValue into different types.

FromRedisKey

A trait used to convert RedisKey values to various types.

HostMappingenable-rustls or enable-native-tls or enable-rustls-ring

A trait used for mapping IP addresses to hostnames when processing the CLUSTER SLOTS response.

ReplicaFilterreplicas

An interface used to filter the list of available replica nodes.

Resolvedns

A trait that can be used to override DNS resolution logic.

Scanner

An interface for interacting with the results of a scan operation.

Type Aliases

Any

The ANY flag used on certain GEO commands.

ConnectHandle

The result from any of the connect functions showing the error that closed the connection, if any.

Limit

A tuple of (offset, count) values for commands that allow paging through results.

LimitCount

An argument type equivalent to “[LIMIT count]”.

MultipleStrings

Convenience interface for commands that take 1 or more strings.

MultipleValues

Convenience interface for commands that take 1 or more values.

Resp2TimeSeriesValuesi-time-series

Shorthand for the result of commands such as MGET, MRANGE, etc.

Resp3TimeSeriesValuesi-time-series

The RESP3 equivalent of Resp2TimeSeriesValues.

XReadResponsei-streams

A generic helper type describing the top level response from XREAD or XREADGROUP.

XReadValuei-streams

A generic helper type describing the ID and associated map for each record in a stream.