Enum fred::types::RedisConfig

pub enum RedisConfig {
    Centralized {
        host: String,
        port: u16,
        key: Option<String>,
        max_value_size: Option<usize>,
    },
    Clustered {
        hosts: Vec<(String, u16)>,
        key: Option<String>,
        max_value_size: Option<usize>,
    },
}

Connection configuration for the Redis server.

Variants
Centralized
Fields of Centralized

host: String

The hostname or IP address of the Redis server.

port: u16

The port on which the Redis server is listening.

key: Option<String>

An optional authentication key to use after connecting.

max_value_size: Option<usize>

The maximum number of bytes that can be allocated for a value, or None for no limit.

Clustered
Fields of Clustered

hosts: Vec<(String, u16)>

A vector of (Host, Port) tuples for nodes in the cluster. Only a subset of nodes in the cluster need to be provided here, the rest will be discovered via the CLUSTER NODES command.

key: Option<String>

An optional authentication key to use after connecting.

max_value_size: Option<usize>

The maximum number of bytes that can be allocated for a value, or None for no limit.

Methods

impl RedisConfig

fn new_centralized(host: String, port: u16, key: Option<String>) -> RedisConfig

fn new_clustered(hosts: Vec<(String, u16)>, key: Option<String>) -> RedisConfig

fn default_centralized() -> RedisConfig

Create a centralized config with default settings for a local deployment.

fn default_clustered() -> RedisConfig

Create a clustered config with the same defaults as specified in the create-cluster script provided by Redis.

fn set_key<T: Into<String>>(&mut self, new_key: Option<T>)

Overwrite the auth key on this config.

fn set_max_size(&mut self, size: Option<usize>)

Set the max size of values received over the socket, or None for no limit.

fn get_max_size(&self) -> Option<usize>

Read a copy of the max_value_size.