Enum fred::types::RedisValue

pub enum RedisValue {
    Boolean(bool),
    Integer(i64),
    Double(f64),
    String(Str),
    Bytes(Bytes),
    Null,
    Queued,
    Map(RedisMap),
    Array(Vec<RedisValue>),
}

A value used in a Redis command.

Variants

Boolean(bool)

A boolean value.

Integer(i64)

An integer value.

Double(f64)

A double floating point number.

String(Str)

A string value.

Bytes(Bytes)

A byte array value.

Null

A nil value.

Queued

A special value used to indicate a MULTI block command was received by the server.

Map(RedisMap)

A map of key/value pairs, primarily used in RESP3 mode.

Array(Vec<RedisValue>)

An ordered list of values.

In RESP2 mode the server usually sends map structures as an array of key/value pairs.

Implementations

impl RedisValue

pub fn from_static(b: &'static [u8]) -> Self

Create a new RedisValue::Bytes from a static byte slice without copying.

pub fn from_static_str(s: &'static str) -> Self

Create a new RedisValue::String from a static str without copying.

pub fn new_ok() -> Self

Create a new RedisValue with the OK status.

pub fn is_ok(&self) -> bool

Whether the value is a simple string OK value.

pub fn into_integer(self) -> Result<RedisValue, RedisValue>

Attempt to convert the value into an integer, returning the original string as an error if the parsing fails.

pub fn kind(&self) -> RedisValueKind

Read the type of the value without any associated data.

pub fn is_null(&self) -> bool

Check if the value is null.

pub fn is_integer(&self) -> bool

Check if the value is an integer.

pub fn is_string(&self) -> bool

Check if the value is a string.

pub fn is_bytes(&self) -> bool

Check if the value is an array of bytes.

pub fn is_boolean(&self) -> bool

Whether the value is a boolean value or can be parsed as a boolean value.

pub fn is_double(&self) -> bool

Whether the inner value is a double or can be parsed as a double.

pub fn is_queued(&self) -> bool

Check if the value is a QUEUED response.

pub fn is_aggregate_type(&self) -> bool

Whether the value is an array or map.

pub fn is_map(&self) -> bool

Whether the value is a RedisMap.

See is_maybe_map for a function that also checks for arrays that likely represent a map in RESP2 mode.

pub fn is_maybe_map(&self) -> bool

Whether the value is a RedisMap or an array with an even number of elements where each even-numbered element is not an aggregate type.

RESP2 and RESP3 encode maps differently, and this function can be used to duck-type maps across protocol versions.

pub fn is_array(&self) -> bool

Whether the value is an array.

pub fn as_u64(&self) -> Option<u64>

Read and return the inner value as a u64, if possible.

pub fn as_i64(&self) -> Option<i64>

Read and return the inner value as a i64, if possible.

pub fn as_usize(&self) -> Option<usize>

Read and return the inner value as a usize, if possible.

pub fn as_f64(&self) -> Option<f64>

Read and return the inner value as a f64, if possible.

pub fn into_string(self) -> Option<String>

Read and return the inner String if the value is a string or scalar value.

pub fn into_bytes_str(self) -> Option<Str>

Read and return the inner data as a Str from the bytes crate.

pub fn as_bytes_str(&self) -> Option<Str>

Read the inner value as a Str.

pub fn as_string(&self) -> Option<String>

Read and return the inner String if the value is a string or scalar value.

Note: this will cast integers and doubles to strings.

pub fn as_str(&self) -> Option<Cow<'_, str>>

Read the inner value as a string slice.

Null is returned as "nil" and scalar values are cast to a string.

pub fn as_str_lossy(&self) -> Option<Cow<'_, str>>

Read the inner value as a string, using String::from_utf8_lossy on byte slices.

pub fn as_bytes(&self) -> Option<&[u8]>

Read the inner value as an array of bytes, if possible.

pub fn as_bool(&self) -> Option<bool>

Attempt to convert the value to a bool.

pub fn into_map(self) -> Result<RedisMap, RedisError>

Attempt to convert this value to a Redis map if it’s an array with an even number of elements.

pub fn into_set(self) -> Result<HashSet<RedisValue>, RedisError>

Convert the array value to a set, if possible.

pub fn into_zset_result(self) -> Result<Vec<(RedisValue, f64)>, RedisError>

Convert a RedisValue to Vec<(RedisValue, f64)>, if possible.

pub fn into_array(self) -> Vec<RedisValue>

Convert this value to an array if it’s an array or map.

If the value is not an array or map this returns a single-element array containing the original value.

pub fn into_owned_bytes(self) -> Option<Vec<u8>>

Convert the value to an array of bytes, if possible.

pub fn into_bytes(self) -> Option<Bytes>

Convert the value into a Bytes view.

pub fn array_len(&self) -> Option<usize>

Return the length of the inner array if the value is an array.

pub fn flatten_array_values(self, depth: usize) -> Self

Flatten adjacent nested arrays to the provided depth.

See the XREAD documentation for an example of when this might be useful.

pub fn into_xread_response<K1, I, K2, V>( self, ) -> Result<XReadResponse<K1, I, K2, V>, RedisError>
where K1: FromRedisKey + Hash + Eq, K2: FromRedisKey + Hash + Eq, I: FromRedis, V: FromRedis,

Available on crate feature i-streams only.

A utility function to convert the response from XREAD or XREADGROUP into a type with a less verbose type declaration.

This function supports responses in both RESP2 and RESP3 formats.

See the XREAD (or XREADGROUP) documentation for more information.

pub fn into_xread_value<I, K, V>( self, ) -> Result<Vec<XReadValue<I, K, V>>, RedisError>
where K: FromRedisKey + Hash + Eq, I: FromRedis, V: FromRedis,

Available on crate feature i-streams only.

A utility function to convert the response from XCLAIM, etc into a type with a less verbose type declaration.

This function supports responses in both RESP2 and RESP3 formats.

pub fn into_xautoclaim_values<I, K, V>( self, ) -> Result<(String, Vec<XReadValue<I, K, V>>), RedisError>
where K: FromRedisKey + Hash + Eq, I: FromRedis, V: FromRedis,

Available on crate feature i-streams only.

A utility function to convert the response from XAUTOCLAIM into a type with a less verbose type declaration.

This function supports responses in both RESP2 and RESP3 formats.

Note: the new (as of Redis 7.x) return value containing message PIDs that were deleted from the PEL are dropped. Callers should use xautoclaim instead if this data is needed.

pub fn as_functions(&self, name: &str) -> Result<Vec<Function>, RedisError>

Available on crate feature i-scripts only.

Parse the value as the response from FUNCTION LIST, including only functions with the provided library name.

pub fn as_geo_position(&self) -> Result<Option<GeoPosition>, RedisError>

Available on crate feature i-geo only.

Convert the value into a GeoPosition, if possible.

Null values are returned as None to work more easily with the result of the GEOPOS command.

pub fn into_geo_radius_result( self, withcoord: bool, withdist: bool, withhash: bool, ) -> Result<Vec<GeoRadiusInfo>, RedisError>

Available on crate feature i-geo only.

Parse the value as the response to any of the relevant GEO commands that return an array of GeoRadiusInfo values, such as GEOSEARCH, GEORADIUS`, etc.

pub fn take(&mut self) -> RedisValue

Replace this value with RedisValue::Null, returning the original value.

pub fn convert<R>(self) -> Result<R, RedisError>
where R: FromRedis,

Attempt to convert this value to any value that implements the FromRedis trait.

pub fn can_hash(&self) -> bool

Whether the value can be hashed.

Some use cases require using RedisValue types as keys in a HashMap, etc. Trying to do so with an aggregate type can panic, and this function can be used to more gracefully handle this situation.

pub fn into_json(self) -> Result<Value, RedisError>

Available on crate feature serde-json only.

Convert the value to JSON.