Struct fred::types::Builder

pub struct Builder { /* private fields */ }

A client and pool builder interface.

fn example() -> Result<(), RedisError> {
  // use default values
  let client = Builder::default_centralized().build()?;

  // or initialize from a URL or config
  let config = RedisConfig::from_url("redis://localhost:6379/1")?;
  let mut builder = Builder::from_config(config);
  // or modify values in place (creating defaults if needed)
  builder
    .with_performance_config(|config| {
      config.auto_pipeline = true;
    })
    .with_config(|config| {
      config.version = RespVersion::RESP3;
      config.fail_fast = true;
    })
    .with_connection_config(|config| {
      config.tcp = TcpConfig {
        nodelay: Some(true),
        ..Default::default()
      };
      config.internal_command_timeout = Duration::from_secs(10);
    });
  // or overwrite configuration structs in place
  builder.set_policy(ReconnectPolicy::new_exponential(0, 100, 30_000, 2));
  builder.set_performance_config(PerformanceConfig::default());

  // reuse the builder as needed to create any kind of client
  let client = builder.build()?;
  let pool = builder.build_pool(3)?;
  let subscriber = builder.build_subscriber_client()?;

  // ...

  Ok(())
}

Implementations

impl Builder

pub fn default_centralized() -> Self

Create a new builder instance with default config values for a centralized deployment.

pub fn default_clustered() -> Self

Create a new builder instance with default config values for a clustered deployment.

pub fn from_config(config: RedisConfig) -> Self

Create a new builder instance from the provided client config.

pub fn get_config(&self) -> Option<&RedisConfig>

Read the client config.

pub fn get_policy(&self) -> Option<&ReconnectPolicy>

Read the reconnection policy.

pub fn get_performance_config(&self) -> &PerformanceConfig

Read the performance config.

pub fn get_connection_config(&self) -> &ConnectionConfig

Read the connection config.

pub fn get_sentinel_config(&self) -> Option<&RedisConfig>

Available on crate feature sentinel-client only.

Read the sentinel client config.

pub fn set_config(&mut self, config: RedisConfig) -> &mut Self

Overwrite the client config on the builder.

pub fn set_policy(&mut self, policy: ReconnectPolicy) -> &mut Self

Overwrite the reconnection policy on the builder.

pub fn set_performance_config(&mut self, config: PerformanceConfig) -> &mut Self

Overwrite the performance config on the builder.

pub fn set_connection_config(&mut self, config: ConnectionConfig) -> &mut Self

Overwrite the connection config on the builder.

pub fn set_sentinel_config(&mut self, config: SentinelConfig) -> &mut Self

Available on crate feature sentinel-client only.

Overwrite the sentinel config on the builder.

pub fn with_config<F>(&mut self, func: F) -> &mut Self
where F: FnOnce(&mut RedisConfig),

Modify the client config in place, creating a new one with default centralized values first if needed.

pub fn with_performance_config<F>(&mut self, func: F) -> &mut Self
where F: FnOnce(&mut PerformanceConfig),

Modify the performance config in place, creating a new one with default values first if needed.

pub fn with_connection_config<F>(&mut self, func: F) -> &mut Self
where F: FnOnce(&mut ConnectionConfig),

Modify the connection config in place, creating a new one with default values first if needed.

pub fn with_sentinel_config<F>(&mut self, func: F) -> &mut Self
where F: FnOnce(&mut SentinelConfig),

Available on crate feature sentinel-client only.

Modify the sentinel config in place, creating a new one with default values first if needed.

pub fn build(&self) -> Result<RedisClient, RedisError>

Create a new client.

pub fn build_pool(&self, size: usize) -> Result<RedisPool, RedisError>

Create a new client pool.

pub fn build_subscriber_client(&self) -> Result<SubscriberClient, RedisError>

Available on crate feature subscriber-client only.

Create a new subscriber client.

pub fn build_sentinel_client(&self) -> Result<SentinelClient, RedisError>

Available on crate feature sentinel-client only.

Create a new sentinel client.

This is only necessary if callers need to communicate directly with sentinel nodes. Use a ServerConfig::Sentinel to interact with Redis servers behind a sentinel layer.