Trait fred::types::FromRedis

pub trait FromRedis: Sized {
    // Required method
    fn from_value(value: RedisValue) -> Result<Self, RedisError>;
}

A trait used to convert various forms of RedisValue into different types.

Examples

let foo: usize = RedisValue::String("123".into()).convert()?;
let foo: i64 = RedisValue::String("123".into()).convert()?;
let foo: String = RedisValue::String("123".into()).convert()?;
let foo: Vec<u8> = RedisValue::Bytes(vec![102, 111, 111].into()).convert()?;
let foo: Vec<u8> = RedisValue::String("foo".into()).convert()?;
let foo: Vec<String> = RedisValue::Array(vec!["a".into(), "b".into()]).convert()?;
let foo: HashMap<String, u16> =
  RedisValue::Array(vec!["a".into(), 1.into(), "b".into(), 2.into()]).convert()?;
let foo: (String, i64) = RedisValue::Array(vec!["a".into(), 1.into()]).convert()?;
let foo: Vec<(String, i64)> =
  RedisValue::Array(vec!["a".into(), 1.into(), "b".into(), 2.into()]).convert()?;
// ...

Bulk Values

This interface can also convert single-element vectors to scalar values in certain scenarios. This is often useful with commands that conditionally return bulk values, or where the number of elements in the response depends on the number of arguments (MGET, etc).

For example:

let _: String = RedisValue::Array(vec![]).convert()?; // error
let _: String = RedisValue::Array(vec!["a".into()]).convert()?; // "a"
let _: String = RedisValue::Array(vec!["a".into(), "b".into()]).convert()?; // error
let _: Option<String> = RedisValue::Array(vec![]).convert()?; // None
let _: Option<String> = RedisValue::Array(vec!["a".into()]).convert()?; // Some("a")
let _: Option<String> = RedisValue::Array(vec!["a".into(), "b".into()]).convert()?; // error

The default-nil-types Feature Flag

By default a nil value cannot be converted directly into any of the scalar types (u8, String, Bytes, etc). In practice this often requires callers to use an Option or Vec container with commands that can return nil.

The default-nil-types feature flag can enable some further type conversion branches that treat nil values as default values for the relevant type. For RedisValue::Null these include:

impl FromRedis for String or Str returns an empty string.

impl FromRedis for Bytes or Vec<T> returns an empty array.

impl FromRedis for any integer or float type returns 0

impl FromRedis for bool returns false

impl FromRedis for map or set types return an empty map or set.

Required Methods

fn from_value(value: RedisValue) -> Result<Self, RedisError>

Object Safety

This trait is not object safe.