Struct fred::clients::Transaction

pub struct Transaction { /* private fields */ }

Available on crate feature transactions only.

A cheaply cloneable transaction block.

Implementations

impl Transaction

pub fn id(&self) -> u64

An ID identifying the underlying transaction state.

pub fn reset(&self)

Clear the internal command buffer and watched keys.

pub fn len(&self) -> usize

Read the number of commands queued to run.

pub fn watched_len(&self) -> usize

Read the number of keys to WATCH before the starting the transaction.

pub async fn exec<R>(&self, abort_on_error: bool) -> Result<R, RedisError>
where R: FromRedis,

Executes all previously queued commands in a transaction.

If abort_on_error is true the client will automatically send DISCARD if an error is received from any of the commands prior to EXEC. This does not apply to MOVED or ASK errors, which wll be followed automatically.

https://redis.io/commands/exec

async fn example(client: &RedisClient) -> Result<(), RedisError> {
  let _ = client.mset(vec![("foo", 1), ("bar", 2)]).await?;

  let trx = client.multi();
  let _: () = trx.get("foo").await?; // returns QUEUED
  let _: () = trx.get("bar").await?; // returns QUEUED

  let (foo, bar): (i64, i64) = trx.exec(false).await?;
  assert_eq!((foo, bar), (1, 2));
  Ok(())
}

pub fn watch_before<K>(&self, keys: K)
where K: Into<MultipleKeys>,

Send the WATCH command with the provided keys before starting the transaction.

pub fn hash_slot(&self) -> Option<u16>

Read the hash slot against which this transaction will run, if known.

pub fn cluster_node(&self) -> Option<Server>

Read the server ID against which this transaction will run, if known.