Trait fred::interfaces::SetsInterface

pub trait SetsInterface: ClientLike + Sized {

    // Provided methods
    fn sadd<R, K, V>(
        &self,
        key: K,
        members: V,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey>,
             V: TryInto<MultipleValues>,
             V::Error: Into<RedisError> { ... }
    fn scard<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey> { ... }
    fn sdiff<R, K>(&self, keys: K) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<MultipleKeys> { ... }
    fn sdiffstore<R, D, K>(
        &self,
        dest: D,
        keys: K,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             D: Into<RedisKey>,
             K: Into<MultipleKeys> { ... }
    fn sinter<R, K>(&self, keys: K) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<MultipleKeys> { ... }
    fn sinterstore<R, D, K>(
        &self,
        dest: D,
        keys: K,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             D: Into<RedisKey>,
             K: Into<MultipleKeys> { ... }
    fn sismember<R, K, V>(
        &self,
        key: K,
        member: V,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey>,
             V: TryInto<RedisValue>,
             V::Error: Into<RedisError> { ... }
    fn smismember<R, K, V>(
        &self,
        key: K,
        members: V,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey>,
             V: TryInto<MultipleValues>,
             V::Error: Into<RedisError> { ... }
    fn smembers<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey> { ... }
    fn smove<R, S, D, V>(
        &self,
        source: S,
        dest: D,
        member: V,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             S: Into<RedisKey>,
             D: Into<RedisKey>,
             V: TryInto<RedisValue>,
             V::Error: Into<RedisError> { ... }
    fn spop<R, K>(
        &self,
        key: K,
        count: Option<usize>,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey> { ... }
    fn srandmember<R, K>(
        &self,
        key: K,
        count: Option<usize>,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey> { ... }
    fn srem<R, K, V>(
        &self,
        key: K,
        members: V,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<RedisKey>,
             V: TryInto<MultipleValues>,
             V::Error: Into<RedisError> { ... }
    fn sunion<R, K>(&self, keys: K) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             K: Into<MultipleKeys> { ... }
    fn sunionstore<R, D, K>(
        &self,
        dest: D,
        keys: K,
    ) -> impl Future<Output = RedisResult<R>>
       where R: FromRedis,
             D: Into<RedisKey>,
             K: Into<MultipleKeys> { ... }

}

Available on crate feature i-sets only.

Functions that implement the sets interface.

Provided Methods

fn sadd<R, K, V>( &self, key: K, members: V, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>, V: TryInto<MultipleValues>, V::Error: Into<RedisError>,

Add the specified members to the set stored at key.

https://redis.io/commands/sadd

fn scard<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>,

Returns the set cardinality (number of elements) of the set stored at key.

https://redis.io/commands/scard

fn sdiff<R, K>(&self, keys: K) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<MultipleKeys>,

Returns the members of the set resulting from the difference between the first set and all the successive sets.

https://redis.io/commands/sdiff

fn sdiffstore<R, D, K>( &self, dest: D, keys: K, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, D: Into<RedisKey>, K: Into<MultipleKeys>,

This command is equal to SDIFF, but instead of returning the resulting set, it is stored in destination.

https://redis.io/commands/sdiffstore

fn sinter<R, K>(&self, keys: K) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<MultipleKeys>,

Returns the members of the set resulting from the intersection of all the given sets.

https://redis.io/commands/sinter

fn sinterstore<R, D, K>( &self, dest: D, keys: K, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, D: Into<RedisKey>, K: Into<MultipleKeys>,

This command is equal to SINTER, but instead of returning the resulting set, it is stored in destination.

https://redis.io/commands/sinterstore

fn sismember<R, K, V>( &self, key: K, member: V, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>, V: TryInto<RedisValue>, V::Error: Into<RedisError>,

Returns if member is a member of the set stored at key.

https://redis.io/commands/sismember

fn smismember<R, K, V>( &self, key: K, members: V, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>, V: TryInto<MultipleValues>, V::Error: Into<RedisError>,

Returns whether each member is a member of the set stored at key.

https://redis.io/commands/smismember

fn smembers<R, K>(&self, key: K) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>,

Returns all the members of the set value stored at key.

https://redis.io/commands/smembers

fn smove<R, S, D, V>( &self, source: S, dest: D, member: V, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, S: Into<RedisKey>, D: Into<RedisKey>, V: TryInto<RedisValue>, V::Error: Into<RedisError>,

Move member from the set at source to the set at destination.

https://redis.io/commands/smove

fn spop<R, K>( &self, key: K, count: Option<usize>, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>,

Removes and returns one or more random members from the set value store at key.

https://redis.io/commands/spop

fn srandmember<R, K>( &self, key: K, count: Option<usize>, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>,

When called with just the key argument, return a random element from the set value stored at key.

If the provided count argument is positive, return an array of distinct elements. The array’s length is either count or the set’s cardinality (SCARD), whichever is lower.

https://redis.io/commands/srandmember

fn srem<R, K, V>( &self, key: K, members: V, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<RedisKey>, V: TryInto<MultipleValues>, V::Error: Into<RedisError>,

Remove the specified members from the set stored at key.

https://redis.io/commands/srem

fn sunion<R, K>(&self, keys: K) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, K: Into<MultipleKeys>,

Returns the members of the set resulting from the union of all the given sets.

https://redis.io/commands/sunion

fn sunionstore<R, D, K>( &self, dest: D, keys: K, ) -> impl Future<Output = RedisResult<R>>
where R: FromRedis, D: Into<RedisKey>, K: Into<MultipleKeys>,

This command is equal to SUNION, but instead of returning the resulting set, it is stored in destination.

https://redis.io/commands/sunionstore

Object Safety

This trait is not object safe.