Module fred::types

Configuration options, return value types, etc.

Structs

MultipleKeys
Convenience struct for commands that take 1 or more keys.

RedisKey
A key in Redis.

Enums

ClientState
The state of the underlying connection to the Redis server.

Expiration
Expiration options for the set command.

InfoKind
Options for the info command.

ReconnectPolicy
The type of reconnection policy to use. This will apply to every connection used by the client.

RedisConfig
Connection configuration for the Redis server.

RedisValue
A value used in a Redis command.

RedisValueKind
The kind of value from Redis.

SetOptions
Options for the set command.

Statics

ASYNC

Type Definitions

ConnectionFuture
A future that resolves when the connection to the Redis server closes.