Enum fred::types::RedisValue

pub enum RedisValue {
    Integer(i64),
    String(String),
    Null,
    // some variants omitted
}

A value used in a Redis command.

Variants
Integer(i64)String(String)Null
Methods

impl RedisValue

fn into_integer(self) -> Result<RedisValue, RedisValue>

Returns the original string as an error if the parsing fails, otherwise this consumes the original string.

fn kind(&self) -> RedisValueKind

Check the specific data type used to represent the value.

fn is_null(&self) -> bool

Check if the value is null.

fn is_integer(&self) -> bool

Check if the value is an integer.

fn is_string(&self) -> bool

Check if the value is a string.

fn is_float(&self) -> bool

Check if the inner string value can be coerced to an f64.

fn into_u64(self) -> Option<u64>

Read and return the inner value as a u64, if possible.

fn into_i64(self) -> Option<i64>

Read and return the inner value as a i64, if possible.

fn into_f64(self) -> Option<f64>

Read and return the inner value as a f64, if possible.

fn into_string(self) -> Option<String>

Read and return the inner String if the value is a string or integer.

fn from_u64(d: u64) -> Result<RedisValue, RedisError>

Convert from a u64 to the i64 representation used by Redis. This can fail due to overflow so it is not implemented via the From trait.

fn from_usize(d: usize) -> Result<RedisValue, RedisError>

fn take(&mut self) -> RedisValue

Replace this RedisValue instance with RedisValue::Null, returning the original value.