Module fred::util

Various client utility functions.

Constants

NONE

A convenience constant for None values used as generic arguments.

Functions

f64_to_redis_string

Convert an f64 to a redis string, supporting “+inf” and “-inf”.

group_by_hash_slot

Group the provided arguments by their cluster hash slot.

redis_keyslot

Map a key to the corresponding cluster key slot.

redis_string_to_f64

Convert a redis string to an f64, supporting “+inf” and “-inf”.

sha1_hashsha-1

Calculate the SHA1 hash output as a hex string. This is provided for clients that use the Lua interface to manage their own script caches.

static_bytes

Create a Bytes from static bytes without copying.

static_str

Create a Str from a static str slice without copying.